Separate the represented movement and world-transfer responsibility (#603)
Fourth delivery of the refactor push. 137 methods move into 7 bounded
submodules under session/movement: state, movement_publication,
movement_validation, speed, fall, transfer and far_transfer. Every file is
under the 1,000-line review budget.

session/mod.rs drops from 54,029 to 50,802 production lines. Across #597,
#599, #601 and #603 the root falls from 75,841 to 50,802 - a third retired
into navigable owners.

The far-transfer writer fence, its cancellation and its recovery ordering,
established by #585 and #586, are relocated into their own module and are
otherwise untouched: this delivery moves the methods that implement them and
changes none of their sequence. 14 methods that mention movement or position
but belong to creature, gameobject, pet, quest, item, loot, spell and aura -
all already separated by earlier deliveries - stay where they are.

Safety established by comparison: the impl surface is 1,714 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged. All 34 baseline differences are private to
pub(in crate::session); nothing became public. The six direct-registry deltas
cover the identical method set, re-attributed to their new files.

The submodule holding general movement state is named state.rs rather than
movement.rs, so the path does not read session::movement::movement.

Verified: cargo check -p wow-world --lib --all-targets clean, cargo test
-p wow-world --lib 3,822 passed and 0 failed, session-ownership-check
--syntax-only PASS, check_architecture.py check and self-test PASS, fmt and
diff hygiene clean. Ceiling tightened to the validated figure.
