+++ crates/wow-world/src/session/mod.rs
@@ -17,6 +17,7 @@ mod lifecycle;
 pub use lifecycle::PlayerSaveOutcomeLikeCpp;
 mod effect_learning;
 pub mod mailbox;
+mod movement;
 mod player_cast;
 mod player_items;
 pub mod registry;
@@ -10165,50 +10166,6 @@ impl WorldSession {
         })
     }
 
-    fn pending_teleport_save_destination_like_cpp(&self) -> Option<(u16, Position)> {
-        if let Some((map_id, position)) = self.pending_teleport_like_cpp() {
-            return Some((u16::try_from(map_id).unwrap_or(u16::MAX), position));
-        }
-        if let Some(teleport) = self.player_teleport_state_snapshot_like_cpp()
-            && teleport.near_pending
-        {
-            return teleport.near_destination;
-        }
-        None
-    }
-
-    pub(crate) fn remove_current_player_from_canonical_current_map_like_cpp(&mut self) -> bool {
-        let Some(guid) = self.player_guid() else {
-            return false;
-        };
-        let Some(manager) = self.canonical_map_manager.as_ref().map(Arc::clone) else {
-            return false;
-        };
-        let Ok(mut manager) = manager.lock() else {
-            return false;
-        };
-        let handle = match self.player_handle_like_cpp {
-            Some(handle) if handle.guid() == guid => handle,
-            Some(_) => return false,
-            None => {
-                // Adopt a unique legacy record before removing it. Discarding
-                // RemoveFromMap's returned Box would destroy the live Player.
-                let Ok(handle) = manager.adopt_active_player_like_cpp(guid) else {
-                    return false;
-                };
-                self.player_handle_like_cpp = Some(handle);
-                handle
-            }
-        };
-        match manager.player_residence_like_cpp(handle) {
-            Some(wow_map::PlayerResidenceLikeCpp::Active(_)) => {
-                manager.detach_player_like_cpp(handle).is_ok()
-            }
-            Some(wow_map::PlayerResidenceLikeCpp::Detached) => true,
-            None => false,
-        }
-    }
-
     pub(crate) fn mutate_canonical_player_like_cpp<R>(
         &self,
         f: impl FnOnce(&mut Player) -> R,
@@ -10963,37 +10920,6 @@ impl WorldSession {
         result.unwrap_or(ObjectGuid::EMPTY)
     }
 
-    pub(crate) fn represented_move_dismiss_vehicle_like_cpp(
-        &mut self,
-        status: &mut wow_packet::packets::movement::MovementInfo,
-    ) -> bool {
-        let vehicle_guid = self.represented_player_charmed_guid_like_cpp();
-        if vehicle_guid.is_empty() || self.player_vehicle_seat_state_like_cpp().is_none() {
-            return false;
-        }
-
-        self.sanitize_movement_info_represented_like_cpp(status);
-        self.set_player_movement_time_like_cpp(status.time);
-        self.set_player_movement_flags_like_cpp(status.flags);
-        self.set_player_position_like_cpp(status.position);
-        self.update_registry_position();
-        #[cfg(test)]
-        self.represented_vehicle_dismiss_movements_like_cpp.push(
-            RepresentedVehicleDismissMovementLikeCpp {
-                vehicle_guid,
-                sanitized_flags: status.flags,
-                position: status.position,
-                time: status.time,
-            },
-        );
-
-        if !self.set_player_vehicle_seat_state_like_cpp(None, None) {
-            return false;
-        }
-        self.sync_player_registry_state_like_cpp();
-        true
-    }
-
     fn attack_reputation_faction_snapshot_like_cpp(
         &self,
         creature: &wow_entities::Creature,
@@ -11662,20 +11588,6 @@ impl WorldSession {
         });
     }
 
-    fn remove_canonical_attacker_like_cpp(&mut self, victim: ObjectGuid, attacker: ObjectGuid) {
-        if self
-            .mutate_canonical_player_by_guid_like_cpp(victim, |victim| {
-                victim.unit_mut().remove_attacker_like_cpp(attacker)
-            })
-            .is_some()
-        {
-            return;
-        }
-        let _ = self.mutate_canonical_creature_by_guid_like_cpp(victim, |victim| {
-            victim.unit_mut().remove_attacker_like_cpp(attacker)
-        });
-    }
-
     fn begin_canonical_player_combat_ref_like_cpp(
         &mut self,
         attacker_guid: ObjectGuid,
@@ -12112,25 +12024,6 @@ impl WorldSession {
         Some(Box::new(
             self.build_initial_player_for_owner_like_cpp(key, None)?,
         ))
-    }
-
-    fn send_transfer_aborted_like_cpp(&self, map_id: u32, transfer_abort: u32) {
-        self.send_transfer_aborted_with_params_like_cpp(map_id, transfer_abort, 0, 0);
-    }
-
-    fn send_transfer_aborted_with_params_like_cpp(
-        &self,
-        map_id: u32,
-        transfer_abort: u32,
-        arg: u8,
-        map_difficulty_x_condition_id: i32,
-    ) {
-        self.send_packet(&wow_packet::packets::misc::TransferAborted {
-            map_id,
-            arg,
-            map_difficulty_x_condition_id,
-            transfer_abort,
-        });
     }
 
     fn prune_expired_instance_reset_times_like_cpp(&mut self, now_secs: u64) {
@@ -14981,41 +14874,6 @@ impl WorldSession {
         Ok(Some(delta))
     }
 
-    /// C++ `Player::RemoveCurrency` underflow guard for vendor costs.
-    pub(crate) fn plan_remove_currency_like_cpp(
-        currencies: &mut HashMap<u32, PlayerCurrency>,
-        currency_id: u32,
-        amount: u32,
-    ) -> bool {
-        if amount == 0 {
-            return true;
-        }
-
-        let Some(currency) = currencies.get_mut(&currency_id) else {
-            return false;
-        };
-        if currency.quantity == 0 {
-            return false;
-        }
-
-        let removed = amount.min(currency.quantity);
-        currency.quantity -= removed;
-        if currency.state != PlayerCurrencyState::New {
-            currency.state = PlayerCurrencyState::Changed;
-        }
-        true
-    }
-
-    pub(crate) fn remove_currency(&mut self, currency_id: u32, amount: u32) -> bool {
-        let Some(mut currencies) = self.player_currencies_like_cpp() else {
-            return false;
-        };
-        if !Self::plan_remove_currency_like_cpp(&mut currencies, currency_id, amount) {
-            return false;
-        }
-        self.set_player_currencies_like_cpp(currencies)
-    }
-
     /// C++ `Player::_SaveCurrency` plan for changed/new currency rows.
     /// Gameplay owns filtering and state transitions; the persistence adapter
     /// owns statement identity, bind order, and transaction execution.
@@ -15696,11 +15554,6 @@ impl WorldSession {
             true
         })
         .unwrap_or(false)
-    }
-
-    pub(crate) fn remove_account_toy_like_cpp(&mut self, item_id: u32) -> bool {
-        self.mutate_player_collection_state_like_cpp(|state| state.toys.remove(&item_id).is_some())
-            .unwrap_or(false)
     }
 
     /// C++ `Player::AddToy`, called from `CollectionMgr::AddToy` after the
@@ -18445,25 +18298,6 @@ impl WorldSession {
         }
     }
 
-    fn position_is_within_area_trigger_box_like_cpp(
-        pos: &Position,
-        center: &Position,
-        half_length: f32,
-        half_width: f32,
-        half_height: f32,
-    ) -> bool {
-        let dx = pos.x - center.x;
-        let dy = pos.y - center.y;
-        let cos_yaw = center.orientation.cos();
-        let sin_yaw = center.orientation.sin();
-        let rel_x = dx * cos_yaw + dy * sin_yaw;
-        let rel_y = -dx * sin_yaw + dy * cos_yaw;
-
-        rel_x.abs() <= half_length
-            && rel_y.abs() <= half_width
-            && (pos.z - center.z).abs() <= half_height
-    }
-
     #[cfg(test)]
     pub fn set_graveyard_store(&mut self, store: Arc<GraveyardStore>) {
         self.graveyard_store = Some(store);
@@ -21150,11 +20984,6 @@ impl WorldSession {
         .unwrap_or(false)
     }
 
-    pub(crate) fn remove_represented_rest_flag_like_cpp(&mut self, rest_flag: u32) -> bool {
-        self.mutate_player_rest_state_like_cpp(|state| state.remove_flag_like_cpp(rest_flag))
-            .unwrap_or(false)
-    }
-
     fn update_represented_rest_flag_like_cpp(&mut self, rest_flag: u32, active: bool) -> bool {
         if active {
             self.set_represented_rest_flag_like_cpp(rest_flag, 0)
@@ -22574,21 +22403,6 @@ impl WorldSession {
         }
 
         Some(wow_packet::packets::misc::TalentInfoLikeCpp { talent_id, rank })
-    }
-
-    fn remove_represented_active_talent_side_effects_like_cpp(&mut self, talent_id: u32, rank: u8) {
-        if let Some(spell_id) = self.represented_talent_spell_id_like_cpp(talent_id, rank) {
-            self.remove_known_spell_like_cpp(spell_id);
-            for trigger_spell in self.represented_direct_learn_spell_triggers_like_cpp(spell_id) {
-                self.remove_known_spell_like_cpp(trigger_spell);
-            }
-        }
-
-        if let Some((overriden_spell_id, new_spell_id)) =
-            self.represented_talent_override_spell_pair_like_cpp(talent_id)
-        {
-            self.remove_represented_override_spell_like_cpp(overriden_spell_id, new_spell_id);
-        }
     }
 
     fn represented_talent_reset_state_plan_like_cpp(
@@ -24927,16 +24741,6 @@ impl WorldSession {
             .expect("test Player emote-state owner must resolve")
     }
 
-    pub(crate) fn clear_player_emote_state_on_movement_like_cpp(
-        &mut self,
-    ) -> Option<wow_packet::packets::update::UpdateObject> {
-        if self.resolved_player_emote_state_like_cpp() == Some(0) {
-            return None;
-        }
-
-        self.set_player_emote_state_like_cpp(0)
-    }
-
     fn player_emote_state_update_packet_like_cpp(
         &self,
         emote_state: u32,
@@ -25077,41 +24881,6 @@ impl WorldSession {
                 .collect(),
         );
         self.sync_player_registry_party_member_party_type_like_cpp();
-    }
-
-    /// Update this session's position (and map) in the player registry.
-    /// Called whenever `player_position` changes.
-    pub(crate) fn update_registry_position(&self) {
-        let (Some(guid), Some(pos), Some(reg)) = (
-            self.player_guid(),
-            self.player_position_like_cpp(),
-            &self.player_registry,
-        ) else {
-            return;
-        };
-        let map_id = self.player_map_id_like_cpp();
-        let Some(is_alive) = self.resolved_player_is_alive_like_cpp() else {
-            return;
-        };
-        // Fallback to 0 (world/default instance) when no canonical map key is
-        // available — mirrors C++ world-map phase where instance_id == 0.
-        let instance_id = self
-            .current_canonical_player_map_key_like_cpp()
-            .map(|k| k.instance_id)
-            .unwrap_or(0);
-        let _ = reg.publish_movement_for_control_channel(
-            guid,
-            &self.session_command_tx,
-            crate::session::directory::PlayerMovementDirectoryUpdate {
-                position: pos,
-                map_id,
-                instance_id,
-                is_in_world: self.player_is_in_world_for_registry_like_cpp(),
-                level: self.player_level_like_cpp(),
-                is_alive,
-                transport: self.player_transport_info_like_cpp(),
-            },
-        );
     }
 
     /// Get the realm ID.
@@ -25433,11 +25202,6 @@ impl WorldSession {
         self.legit_characters.contains(guid)
     }
 
-    /// Remove a GUID from the legit characters list.
-    pub fn remove_legit_character(&mut self, guid: &ObjectGuid) {
-        self.legit_characters.retain(|g| g != guid);
-    }
-
     // ── Aura system ───────────────────────────────────────────────
 
     fn apply_represented_battle_pet_xp_pct_aura_like_cpp(
@@ -25667,84 +25431,6 @@ impl WorldSession {
         }
     }
 
-    fn send_movement_set_collision_height_like_cpp(&mut self, reason: u8) {
-        let Some(player_guid) = self.player_guid() else {
-            return;
-        };
-        let Some((_, mount_display_id, object_scale)) =
-            self.player_unit_presentation_snapshot_like_cpp()
-        else {
-            return;
-        };
-        let Some(collision_height) = self
-            .with_owned_player_like_cpp(|player| player.unit().collision_height_like_cpp())
-            .or_else(|| {
-                #[cfg(test)]
-                {
-                    return self
-                        .player_handle_like_cpp
-                        .is_none()
-                        .then_some(self.player_collision_height_like_cpp);
-                }
-                #[cfg(not(test))]
-                {
-                    None
-                }
-            })
-        else {
-            return;
-        };
-        let Some(sequence_index) = self.next_movement_counter_like_cpp() else {
-            return;
-        };
-
-        let Some(scale_duration) = self.resolved_player_scale_duration_like_cpp() else {
-            return;
-        };
-
-        self.send_packet(&wow_packet::packets::movement::MoveSetCollisionHeight {
-            mover_guid: player_guid,
-            sequence_index,
-            height: collision_height,
-            scale: object_scale,
-            reason,
-            mount_display_id: u32::try_from(mount_display_id).unwrap_or(0),
-            scale_duration,
-        });
-
-        use wow_packet::ServerPacket;
-        let Some(status) = self.current_player_movement_info_like_cpp(player_guid) else {
-            return;
-        };
-        self.broadcast_to_movement_set_like_cpp(
-            wow_packet::packets::movement::MoveUpdateCollisionHeight {
-                status,
-                height: collision_height,
-                scale: object_scale,
-            }
-            .to_bytes(),
-            false,
-        );
-    }
-
-    fn current_player_movement_info_like_cpp(
-        &self,
-        player_guid: ObjectGuid,
-    ) -> Option<wow_packet::packets::movement::MovementInfo> {
-        Some(wow_packet::packets::movement::MovementInfo {
-            guid: player_guid,
-            position: self.player_position_like_cpp()?,
-            flags: self.resolved_player_movement_flags_like_cpp()?,
-            flags2: if self.resolved_can_swim_to_fly_transition_like_cpp()? {
-                wow_constants::movement::MovementFlag2::CAN_SWIM_TO_FLY_TRANS
-            } else {
-                wow_constants::movement::MovementFlag2::NONE
-            },
-            time: self.resolved_player_movement_time_like_cpp()?,
-            ..wow_packet::packets::movement::MovementInfo::default()
-        })
-    }
-
     fn send_represented_mount_unit_update_like_cpp(&mut self, display_id: i32) {
         let Some(player_guid) = self.player_guid() else {
             return;
@@ -25772,39 +25458,6 @@ impl WorldSession {
             SPELL_AURA_INTERRUPT_FLAG_LOOTING_LIKE_CPP,
             0,
         )
-    }
-
-    pub(crate) fn remove_represented_feign_death_if_needed_like_cpp(&mut self) -> bool {
-        let has_died_state = self
-            .mutate_canonical_player_like_cpp(|player| {
-                player.unit().has_unit_state(UnitState::DIED.bits())
-            })
-            .unwrap_or(false);
-        if !has_died_state {
-            return false;
-        }
-
-        let Some(visible_auras) = self.resolved_player_visible_auras_like_cpp() else {
-            return false;
-        };
-        let slots: Vec<u8> = visible_auras
-            .iter()
-            .filter_map(|(slot, aura)| {
-                (aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::FeignDeath))
-                    .then_some(*slot)
-            })
-            .collect();
-        if slots.is_empty() {
-            return false;
-        }
-
-        for slot in slots {
-            let _ = self.remove_aura(slot);
-        }
-        let _ = self.mutate_canonical_player_like_cpp(|player| {
-            player.unit_mut().clear_unit_state(UnitState::DIED.bits());
-        });
-        true
     }
 
     /// Send a TimeSyncRequest and schedule the next one.
@@ -25925,21 +25578,6 @@ impl WorldSession {
         }
     }
 
-    pub(crate) fn adjust_client_movement_time_like_cpp(&self, time: u32) -> u32 {
-        let movement_time = i64::from(time) + self.time_sync_clock_delta;
-        if self.time_sync_clock_delta == 0 || !(0..=i64::from(u32::MAX)).contains(&movement_time) {
-            warn!(
-                account = self.account_id,
-                client_time = time,
-                clock_delta = self.time_sync_clock_delta,
-                "The computed movement time using clockDelta is erroneous. Using fallback instead"
-            );
-            Self::game_time_ms_like_cpp()
-        } else {
-            movement_time as u32
-        }
-    }
-
     /// Check for area triggers at the player's current position.
     ///
     /// This is called after movement updates to handle:
@@ -26031,534 +25669,6 @@ impl WorldSession {
             .await;
     }
 
-    /// Teleport the player to a new map and position.
-    ///
-    /// Sends SMSG_TRANSFER_PENDING (0x25cd) to initiate the transfer.
-    /// The client will respond with CMSG_WORLD_PORT_ACK when ready.
-    ///
-    /// C++ `Player::TeleportTo` → `SendTransferPending`.
-    pub async fn teleport_to(&mut self, new_map: u32, new_pos: wow_core::Position) {
-        self.teleport_to_with_options(new_map, new_pos, TELE_TO_NONE_LIKE_CPP)
-            .await;
-    }
-
-    pub(crate) async fn teleport_to_with_options(
-        &mut self,
-        new_map: u32,
-        new_pos: wow_core::Position,
-        mut options: TeleportToOptionsLikeCpp,
-    ) {
-        if self
-            .player_teleport_state_snapshot_like_cpp()
-            .is_some_and(|state| state.recovery == wow_entities::PlayerTransferRecovery::Terminal)
-        {
-            return;
-        }
-        // C++ Player.cpp:1239 validates the catalog map and all four coordinates
-        // before movement, combat, pet, ownership or teleport state is changed.
-        if self
-            .map_store
-            .as_ref()
-            .is_none_or(|maps| maps.get(new_map).is_none())
-            || !new_pos.is_valid_map_coord_like_cpp()
-        {
-            warn!(
-                "Invalid map or coordinates for teleport to {} from account {}",
-                new_map, self.account_id
-            );
-            return;
-        }
-
-        if self.is_map_disabled_for_player_like_cpp(new_map) {
-            warn!(
-                account = self.account_id,
-                map_id = new_map,
-                "Teleport blocked by C++ DisableMgr map gate"
-            );
-            self.send_transfer_aborted_like_cpp(new_map, TRANSFER_ABORT_MAP_NOT_ALLOWED_LIKE_CPP);
-            return;
-        }
-
-        if let Some(target_map) = self
-            .map_store
-            .as_ref()
-            .and_then(|store| store.get(new_map).copied())
-            && target_map.is_battleground_or_arena()
-            && !self.player_in_represented_battleground_like_cpp()
-        {
-            warn!(
-                account = self.account_id,
-                map_id = new_map,
-                "Teleport silently blocked by C++ battleground assignment gate"
-            );
-            return;
-        }
-
-        if let Some(target_map) = self
-            .map_store
-            .as_ref()
-            .and_then(|store| store.get(new_map).copied())
-            && self.expansion < target_map.expansion_like_cpp()
-        {
-            warn!(
-                account = self.account_id,
-                map_id = new_map,
-                session_expansion = self.expansion,
-                required_expansion = target_map.expansion_like_cpp(),
-                "Teleport blocked by C++ client expansion gate"
-            );
-            self.send_transfer_aborted_with_params_like_cpp(
-                new_map,
-                TRANSFER_ABORT_INSUF_EXPAN_LVL_LIKE_CPP,
-                target_map.expansion_like_cpp(),
-                0,
-            );
-            return;
-        }
-
-        let Some(current_pos) = self.player_position_like_cpp() else {
-            warn!(
-                "Cannot teleport account {}: no current position",
-                self.account_id
-            );
-            return;
-        };
-
-        self.exit_represented_vehicle_for_teleport_like_cpp();
-        self.reset_teleport_movement_state_like_cpp();
-
-        if self.player_class_like_cpp() == CLASS_DEATH_KNIGHT_LIKE_CPP
-            && self.player_map_id_like_cpp() == DEATH_KNIGHT_START_MAP_LIKE_CPP
-            && u32::from(self.player_map_id_like_cpp()) != new_map
-            && self.player_is_game_master_like_cpp() != Some(true)
-            && !self
-                .known_spells_like_cpp()
-                .contains(&DEATH_KNIGHT_ESCAPE_SPELL_LIKE_CPP)
-        {
-            self.send_transfer_aborted_with_params_like_cpp(
-                new_map,
-                TRANSFER_ABORT_UNIQUE_MESSAGE_LIKE_CPP,
-                1,
-                0,
-            );
-            return;
-        }
-
-        // Approved recovery exception: a detached Player needs AddPlayerToMap,
-        // even when returning to its old map. A near ACK only relocates it.
-        let active_map = self.current_canonical_player_map_key_like_cpp();
-        let same_map_near_teleport = active_map.is_some_and(|key| key.map_id == new_map);
-        if same_map_near_teleport {
-            if !self.set_represented_far_teleport_pending_like_cpp(false) {
-                return;
-            }
-            self.initiate_same_map_near_teleport_like_cpp(new_map, new_pos, options);
-            return;
-        }
-
-        if let Some((transfer_abort, arg, map_difficulty_x_condition_id)) =
-            self.player_cannot_enter_target_map_like_cpp(new_map)
-        {
-            self.send_transfer_aborted_with_params_like_cpp(
-                new_map,
-                transfer_abort,
-                arg,
-                map_difficulty_x_condition_id,
-            );
-            return;
-        }
-
-        options = self.teleport_options_after_seamless_gate_like_cpp(new_map, options);
-        if active_map.is_none() {
-            options &= !TELE_TO_SEAMLESS_LIKE_CPP;
-        }
-
-        if !same_map_near_teleport {
-            let Some(can_delay) = self
-                .player_teleport_state_snapshot_like_cpp()
-                .map(|state| state.can_delay)
-            else {
-                return;
-            };
-            if !self.update_player_teleport_state_like_cpp(|state| {
-                state.has_delayed = can_delay;
-                state.near_pending = false;
-                state.near_destination = None;
-                state.near_destination_zone_area = None;
-                if can_delay {
-                    state.far_pending = true;
-                    state.delayed = Some((new_map, new_pos, options));
-                }
-            }) {
-                return;
-            }
-            if can_delay {
-                return;
-            }
-
-            self.set_selection_guid_like_cpp(None);
-            self.combat_stop_like_cpp();
-            self.reset_contested_pvp_like_cpp();
-            self.maybe_leave_represented_battleground_on_far_teleport_like_cpp(new_map);
-            self.unsummon_represented_pet_temporary_if_any_like_cpp();
-            let _ = self.remove_all_dynamic_objects_for_current_player_like_cpp();
-            let _ = self.remove_all_area_triggers_for_current_player_like_cpp();
-            if options & TELE_TO_SPELL_LIKE_CPP == 0 {
-                let _ = self.interrupt_non_melee_spells_for_far_teleport_like_cpp();
-            }
-            let _ = self.remove_moving_or_turning_interrupt_auras_for_far_teleport_like_cpp();
-        }
-
-        info!(
-            account = self.account_id,
-            old_map = self.player_map_id_like_cpp(),
-            new_map = new_map,
-            old_pos = format!(
-                "({:.2}, {:.2}, {:.2})",
-                current_pos.x, current_pos.y, current_pos.z
-            ),
-            new_pos = format!("({:.2}, {:.2}, {:.2})", new_pos.x, new_pos.y, new_pos.z),
-            "Player teleporting to new map"
-        );
-
-        use wow_packet::packets::misc::{SuspendToken, TransferPending};
-
-        // 1. SMSG_TRANSFER_PENDING — tell client to start loading screen
-        if !self.player_logout_like_cpp && options & TELE_TO_SEAMLESS_LIKE_CPP == 0 {
-            let transfer_pending = TransferPending {
-                map_id: new_map,
-                old_map_position: current_pos,
-                ship: None,
-                transfer_spell_id: None,
-            };
-            self.send_packet_realm(&transfer_pending);
-            self.clear_active_player_transport_server_time_override_for_far_teleport_like_cpp();
-        }
-
-        let _ = self.remove_current_player_from_canonical_current_map_like_cpp();
-
-        // 2. Store pending destination — completed in handle_world_port_response
-        if !self.set_pending_teleport_like_cpp(Some((new_map, new_pos))) {
-            return;
-        }
-        self.active_area_trigger = None;
-
-        // Retain native completion authority before an interruptible writer wait.
-        if !self.set_represented_far_teleport_pending_like_cpp(true) {
-            return;
-        }
-        self.state = SessionState::Transfer;
-
-        // 3. SMSG_SUSPEND_TOKEN — pause movement processing on client. C++
-        // Player::TeleportTo sets SequenceIndex = m_movementCounter WITHOUT incrementing
-        // (Player.cpp:1466), and HandleMoveWorldportAck's ResumeToken uses the SAME counter
-        // (the before-add reset happens only after ResumeToken). The client pairs suspend and
-        // resume by this index, so they MUST match — a hardcoded 1 here vs the real counter in
-        // ResumeToken left the client stuck on the loading screen. #NEXT.R8.ENTITIES.1229.
-        if !self.player_logout_like_cpp {
-            if options & TELE_TO_SEAMLESS_LIKE_CPP == 0
-                && !self
-                    .wait_for_realm_send_before_instance_update_like_cpp()
-                    .await
-            {
-                self.kick("TransferPending writer fence failed; retain native destination");
-                return;
-            }
-            let Some(suspend_seq) = self.movement_counter_like_cpp() else {
-                return;
-            };
-            self.send_packet(&SuspendToken {
-                sequence_index: suspend_seq,
-                reason: if options & TELE_TO_SEAMLESS_LIKE_CPP != 0 {
-                    2
-                } else {
-                    1
-                },
-            });
-        }
-
-        info!(
-            account = self.account_id,
-            "Teleport initiated: map {} → {} dest ({:.2}, {:.2}, {:.2}); awaiting WorldPortResponse",
-            self.player_map_id_like_cpp(),
-            new_map,
-            new_pos.x,
-            new_pos.y,
-            new_pos.z
-        );
-    }
-
-    fn exit_represented_vehicle_for_teleport_like_cpp(&mut self) -> bool {
-        if self
-            .player_vehicle_seat_state_like_cpp()
-            .and_then(|(flags, _)| flags)
-            .is_none()
-        {
-            return false;
-        }
-
-        if !self.set_player_vehicle_seat_state_like_cpp(None, None) {
-            return false;
-        }
-        self.sync_player_registry_state_like_cpp();
-        true
-    }
-
-    fn reset_teleport_movement_state_like_cpp(&mut self) {
-        let Some(movement_flags) = self.resolved_player_movement_flags_like_cpp() else {
-            return;
-        };
-        let movement_flags = movement_flags & MovementFlag::MASK_HAS_PLAYER_STATUS_OPCODE;
-        self.set_player_movement_flags_like_cpp(movement_flags);
-        #[cfg(test)]
-        {
-            self.player_movement_jump_like_cpp = wow_packet::packets::movement::JumpInfo::default();
-        }
-        let _ = self.mutate_canonical_player_like_cpp(|player| {
-            let motion = &mut player.unit_mut().subsystems_mut().motion;
-            motion.interrupt_spline();
-            let _ =
-                motion.remove_generator_kind(MovementGeneratorKind::Effect, MovementSlot::Active);
-        });
-    }
-
-    fn teleport_options_after_seamless_gate_like_cpp(
-        &self,
-        new_map: u32,
-        mut options: TeleportToOptionsLikeCpp,
-    ) -> TeleportToOptionsLikeCpp {
-        if options & TELE_TO_SEAMLESS_LIKE_CPP == 0 {
-            return options;
-        }
-
-        let Some(map_store) = self.map_store.as_ref() else {
-            return options & !TELE_TO_SEAMLESS_LIKE_CPP;
-        };
-        let Some(old_map_entry) = map_store
-            .get(u32::from(self.player_map_id_like_cpp()))
-            .copied()
-        else {
-            return options & !TELE_TO_SEAMLESS_LIKE_CPP;
-        };
-        let Some(new_map_entry) = map_store.get(new_map).copied() else {
-            return options & !TELE_TO_SEAMLESS_LIKE_CPP;
-        };
-
-        let old_cosmetic_parent = i32::from(old_map_entry.cosmetic_parent_map_id);
-        let new_cosmetic_parent = i32::from(new_map_entry.cosmetic_parent_map_id);
-        let current_map_id = i32::from(self.player_map_id_like_cpp());
-        let new_map_id = i32::try_from(new_map).unwrap_or(i32::MAX);
-        if old_cosmetic_parent != new_map_id
-            && current_map_id != new_cosmetic_parent
-            && !((old_cosmetic_parent != -1) ^ (old_cosmetic_parent != new_cosmetic_parent))
-        {
-            options &= !TELE_TO_SEAMLESS_LIKE_CPP;
-        }
-
-        options
-    }
-
-    fn initiate_same_map_near_teleport_like_cpp(
-        &mut self,
-        map_id: u32,
-        destination: wow_core::Position,
-        options: TeleportToOptionsLikeCpp,
-    ) {
-        let Some(can_delay) = self
-            .player_teleport_state_snapshot_like_cpp()
-            .map(|state| state.can_delay)
-        else {
-            return;
-        };
-        if !self.update_player_teleport_state_like_cpp(|state| {
-            state.has_delayed = can_delay;
-        }) {
-            return;
-        }
-        if can_delay {
-            let map_id_u16 = u16::try_from(map_id).unwrap_or(self.player_map_id_like_cpp());
-            let _ = self.update_player_teleport_state_like_cpp(|state| {
-                state.near_pending = true;
-                state.near_destination = Some((map_id_u16, destination));
-                state.near_destination_zone_area = None;
-                state.delayed = Some((map_id, destination, options));
-            });
-            return;
-        }
-
-        self.unsummon_represented_pet_for_same_map_teleport_if_out_of_range_like_cpp(
-            destination,
-            options,
-        );
-
-        if self.resolved_player_is_alive_like_cpp() == Some(false)
-            && options & TELE_REVIVE_AT_TELEPORT_LIKE_CPP != 0
-        {
-            self.resurrect_player_percent_for_teleport_like_cpp(0.5);
-        }
-
-        if options & TELE_TO_NOT_LEAVE_COMBAT_LIKE_CPP == 0 {
-            self.combat_stop_like_cpp();
-        }
-
-        let map_id = u16::try_from(map_id).unwrap_or(self.player_map_id_like_cpp());
-        if !self.update_player_teleport_state_like_cpp(|state| {
-            state.far_destination = None;
-            state.near_pending = true;
-            state.near_destination = Some((map_id, destination));
-            state.near_destination_zone_area = None;
-        }) {
-            return;
-        }
-        if let Some(current_pos) = self.player_position_like_cpp() {
-            self.set_fall_information_like_cpp(0, current_pos.z);
-        }
-
-        if !self.player_logout_like_cpp {
-            let Some(sequence_index) = self.next_movement_counter_like_cpp() else {
-                return;
-            };
-            if let Some(mover_guid) = self.player_guid() {
-                self.send_same_map_move_update_teleport_to_visible_set_like_cpp(mover_guid);
-                self.send_packet(&wow_packet::packets::movement::MoveTeleport {
-                    mover_guid,
-                    position: destination,
-                    facing: destination.orientation,
-                    sequence_index,
-                    preload_world: 0,
-                    transport_guid: None,
-                });
-            }
-        }
-
-        info!(
-            account = self.account_id,
-            map_id,
-            new_pos = format!(
-                "({:.2}, {:.2}, {:.2})",
-                destination.x, destination.y, destination.z
-            ),
-            "Player same-map near teleport initiated; awaiting MoveTeleportAck"
-        );
-    }
-
-    async fn process_represented_delayed_teleport_after_update_like_cpp(&mut self) -> bool {
-        let Some(teleport) = self.player_teleport_state_snapshot_like_cpp() else {
-            return false;
-        };
-        if !teleport.has_delayed || self.resolved_player_is_alive_like_cpp() != Some(true) {
-            return false;
-        }
-
-        let Some((map_id, destination, options)) = teleport.delayed else {
-            let _ = self.update_player_teleport_state_like_cpp(|state| {
-                state.has_delayed = false;
-            });
-            return false;
-        };
-
-        if !self.update_player_teleport_state_like_cpp(|state| {
-            state.delayed = None;
-            state.can_delay = false;
-            state.has_delayed = false;
-        }) {
-            return false;
-        }
-        if self
-            .current_canonical_player_map_key_like_cpp()
-            .is_some_and(|key| key.map_id == map_id)
-        {
-            self.initiate_same_map_near_teleport_like_cpp(map_id, destination, options);
-        } else {
-            self.initiate_far_teleport_after_delay_like_cpp(map_id, destination, options)
-                .await;
-        }
-        true
-    }
-
-    async fn initiate_far_teleport_after_delay_like_cpp(
-        &mut self,
-        map_id: u32,
-        destination: wow_core::Position,
-        options: TeleportToOptionsLikeCpp,
-    ) {
-        if !self.update_player_teleport_state_like_cpp(|state| {
-            state.near_pending = false;
-            state.near_destination = None;
-            state.near_destination_zone_area = None;
-        }) {
-            return;
-        }
-        let options = if self.current_canonical_player_map_key_like_cpp().is_none() {
-            options & !TELE_TO_SEAMLESS_LIKE_CPP
-        } else {
-            options
-        };
-        self.set_selection_guid_like_cpp(None);
-        self.combat_stop_like_cpp();
-        self.reset_contested_pvp_like_cpp();
-        self.maybe_leave_represented_battleground_on_far_teleport_like_cpp(map_id);
-        self.unsummon_represented_pet_temporary_if_any_like_cpp();
-        let _ = self.remove_all_dynamic_objects_for_current_player_like_cpp();
-        let _ = self.remove_all_area_triggers_for_current_player_like_cpp();
-        if options & TELE_TO_SPELL_LIKE_CPP == 0 {
-            let _ = self.interrupt_non_melee_spells_for_far_teleport_like_cpp();
-        }
-        let _ = self.remove_moving_or_turning_interrupt_auras_for_far_teleport_like_cpp();
-
-        let Some(current_pos) = self.player_position_like_cpp() else {
-            return;
-        };
-
-        use wow_packet::packets::misc::{SuspendToken, TransferPending};
-
-        if !self.player_logout_like_cpp && options & TELE_TO_SEAMLESS_LIKE_CPP == 0 {
-            let transfer_pending = TransferPending {
-                map_id,
-                old_map_position: current_pos,
-                ship: None,
-                transfer_spell_id: None,
-            };
-            self.send_packet_realm(&transfer_pending);
-            self.clear_active_player_transport_server_time_override_for_far_teleport_like_cpp();
-        }
-
-        let _ = self.remove_current_player_from_canonical_current_map_like_cpp();
-
-        if !self.set_pending_teleport_like_cpp(Some((map_id, destination))) {
-            return;
-        }
-        self.active_area_trigger = None;
-        if !self.set_represented_far_teleport_pending_like_cpp(true) {
-            return;
-        }
-        self.state = SessionState::Transfer;
-
-        if !self.player_logout_like_cpp {
-            if options & TELE_TO_SEAMLESS_LIKE_CPP == 0
-                && !self
-                    .wait_for_realm_send_before_instance_update_like_cpp()
-                    .await
-            {
-                self.kick("Delayed TransferPending writer fence failed; retain native destination");
-                return;
-            }
-            // C++ SuspendToken.SequenceIndex = m_movementCounter (Player.cpp:1466); must match
-            // the ResumeToken sent later so the client resumes. #NEXT.R8.ENTITIES.1229.
-            let Some(suspend_seq) = self.movement_counter_like_cpp() else {
-                return;
-            };
-            self.send_packet(&SuspendToken {
-                sequence_index: suspend_seq,
-                reason: if options & TELE_TO_SEAMLESS_LIKE_CPP != 0 {
-                    2
-                } else {
-                    1
-                },
-            });
-        }
-    }
-
     fn unsummon_represented_pet_for_same_map_teleport_if_out_of_range_like_cpp(
         &mut self,
         destination: wow_core::Position,
@@ -26596,109 +25706,6 @@ impl WorldSession {
         if should_unsummon {
             self.unsummon_represented_pet_temporary_if_any_like_cpp();
         }
-    }
-
-    fn resurrect_player_percent_for_teleport_like_cpp(&mut self, restore_percent: f32) {
-        let restored = self.with_owned_player_mut_like_cpp(|player| {
-            let max_health = player
-                .unit()
-                .data()
-                .max_health
-                .clamp(1, u64::from(u32::MAX)) as u32;
-            let health = ((max_health as f32) * restore_percent)
-                .max(0.0)
-                .min(max_health as f32) as u32;
-            player
-                .unit_mut()
-                .set_death_state(wow_constants::DeathState::Alive);
-            player.unit_mut().set_health(u64::from(health));
-            let mana = ((player.get_max_power(PowerType::Mana).max(0) as f32) * restore_percent)
-                .max(0.0) as i32;
-            let energy = ((player.get_max_power(PowerType::Energy).max(0) as f32) * restore_percent)
-                .max(0.0) as i32;
-            let focus = ((player.get_max_power(PowerType::Focus).max(0) as f32) * restore_percent)
-                .max(0.0) as i32;
-            player.unit_mut().set_power(PowerType::Mana, mana);
-            player.unit_mut().set_power(PowerType::Rage, 0);
-            player.unit_mut().set_power(PowerType::Energy, energy);
-            player.unit_mut().set_power(PowerType::Focus, focus);
-            health
-        });
-        if restored.is_none() {
-            return;
-        }
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    fn send_same_map_move_update_teleport_to_visible_set_like_cpp(&self, source_guid: ObjectGuid) {
-        use wow_packet::ServerPacket;
-
-        let Some(registry) = self.player_registry() else {
-            return;
-        };
-        let Some(source_position) = self.player_position_like_cpp() else {
-            return;
-        };
-        let map_id = self.player_map_id_like_cpp();
-        let instance_id = self
-            .current_canonical_player_map_key_like_cpp()
-            .map(|key| key.instance_id)
-            .unwrap_or(0);
-        let Some(status) = self.current_player_movement_info_like_cpp(source_guid) else {
-            return;
-        };
-        let packet_bytes = wow_packet::packets::movement::MoveUpdateTeleport { status }.to_bytes();
-
-        for registration in registry.movement_recipients_within_range(
-            source_guid,
-            map_id,
-            instance_id,
-            source_position,
-            crate::map_manager::VISIBILITY_RADIUS,
-        ) {
-            let _ = registry.try_send_current_command(
-                registration,
-                crate::session::mailbox::SessionCommand::SendIfVisibleLikeCpp(
-                    crate::session::mailbox::SendIfVisibleLikeCppCommand {
-                        queued_at: Instant::now(),
-                        source_guid,
-                        map_id,
-                        instance_id,
-                        packet_bytes: packet_bytes.clone(),
-                    },
-                ),
-            );
-        }
-    }
-
-    fn remove_all_dynamic_objects_for_current_player_like_cpp(
-        &self,
-    ) -> Option<wow_map::map::RemoveAllDynamicObjectsForCasterOutcomeLikeCpp> {
-        let player_guid = self.player_guid()?;
-        let map_key = self.current_canonical_player_map_key_like_cpp()?;
-        let canonical = self.canonical_map_manager.as_ref()?;
-        let mut manager = canonical.lock().ok()?;
-        let managed = manager.find_map_mut(map_key.map_id, map_key.instance_id)?;
-        Some(
-            managed
-                .map_mut()
-                .remove_all_dynamic_objects_for_caster_like_cpp(player_guid),
-        )
-    }
-
-    fn remove_all_area_triggers_for_current_player_like_cpp(
-        &self,
-    ) -> Option<wow_map::map::RemoveAllAreaTriggersForCasterOutcomeLikeCpp> {
-        let player_guid = self.player_guid()?;
-        let map_key = self.current_canonical_player_map_key_like_cpp()?;
-        let canonical = self.canonical_map_manager.as_ref()?;
-        let mut manager = canonical.lock().ok()?;
-        let managed = manager.find_map_mut(map_key.map_id, map_key.instance_id)?;
-        Some(
-            managed
-                .map_mut()
-                .remove_all_area_triggers_for_caster_like_cpp(player_guid),
-        )
     }
 
     fn player_cannot_enter_target_map_like_cpp(&self, map_id: u32) -> Option<(u32, u8, i32)> {
@@ -27832,108 +26839,6 @@ impl WorldSession {
         }
     }
 
-    pub(crate) fn set_player_map_position_like_cpp(
-        &mut self,
-        map_id: u16,
-        position: wow_core::Position,
-    ) {
-        if self.player_map_id_like_cpp() != map_id {
-            self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-        }
-        self.current_map_id = map_id;
-        self.sync_canonical_player_position_if_same_or_detached_like_cpp(map_id, position);
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none()
-            || self.player_position_like_cpp() == Some(position)
-        {
-            self.player_position = Some(position);
-        }
-    }
-
-    fn sync_canonical_player_position_if_same_or_detached_like_cpp(
-        &mut self,
-        map_id: u16,
-        position: Position,
-    ) {
-        let (Some(manager), Some(handle)) = (
-            self.canonical_map_manager.as_ref().map(Arc::clone),
-            self.player_handle_like_cpp,
-        ) else {
-            return;
-        };
-        let Ok(mut manager) = manager.lock() else {
-            return;
-        };
-        match manager.player_residence_like_cpp(handle) {
-            Some(wow_map::PlayerResidenceLikeCpp::Detached) => {
-                let _ = manager.with_player_mut_like_cpp(handle, |player| {
-                    player.unit_mut().world_mut().relocate(position);
-                });
-            }
-            Some(wow_map::PlayerResidenceLikeCpp::Active(key))
-                if key.map_id == u32::from(map_id) =>
-            {
-                let _ = manager.relocate_player_like_cpp(handle, position);
-            }
-            Some(wow_map::PlayerResidenceLikeCpp::Active(_)) | None => {}
-        }
-    }
-
-    pub(crate) fn set_player_position_like_cpp(&mut self, position: wow_core::Position) {
-        self.set_player_map_position_like_cpp(self.current_map_id, position);
-    }
-
-    pub(crate) fn set_player_movement_time_like_cpp(&mut self, time: u32) {
-        let _canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.unit_mut().set_movement_time_like_cpp(time);
-            })
-            .is_some();
-        #[cfg(test)]
-        if _canonical || self.player_handle_like_cpp.is_none() {
-            self.player_movement_time_like_cpp = time;
-        }
-    }
-
-    pub(crate) fn set_player_movement_flags_like_cpp(&mut self, flags: MovementFlag) {
-        let _canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.unit_mut().set_movement_flags_like_cpp(flags);
-            })
-            .is_some();
-        #[cfg(test)]
-        if _canonical || self.player_handle_like_cpp.is_none() {
-            self.player_movement_flags_like_cpp = flags;
-        }
-    }
-
-    pub(crate) fn set_player_movement_jump_like_cpp(
-        &mut self,
-        jump: wow_packet::packets::movement::JumpInfo,
-    ) {
-        #[cfg(test)]
-        {
-            self.player_movement_jump_like_cpp = jump;
-        }
-        #[cfg(not(test))]
-        let _ = jump;
-    }
-
-    pub(crate) fn set_represented_mover_fixed_position_vehicle_like_cpp(&mut self, fixed: bool) {
-        let _canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player
-                    .gameplay_state_mut()
-                    .movement_control
-                    .mover_fixed_position_vehicle = fixed;
-            })
-            .is_some();
-        #[cfg(test)]
-        if _canonical || self.player_handle_like_cpp.is_none() {
-            self.represented_mover_fixed_position_vehicle_like_cpp = fixed;
-        }
-    }
-
     pub(crate) fn set_player_liquid_status_like_cpp(&mut self, status: u32) {
         crate::canonical_player_sync::sync_player_liquid_status_like_cpp(self, status);
     }
@@ -28969,50 +27874,6 @@ impl WorldSession {
             .collect()
     }
 
-    /// C++ `Unit::m_movementCounter` post-increment: returns the current value and advances
-    /// it. Used as the SequenceIndex of movement-control packets (vehicle-rec, collision,
-    /// near-teleport, speed/flag) and read for `SMSG_RESUME_TOKEN` on far teleport.
-    pub(crate) fn next_movement_counter_like_cpp(&mut self) -> Option<u32> {
-        let canonical = self.with_owned_player_mut_like_cpp(|player| {
-            player.unit_mut().next_movement_counter_like_cpp()
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            let sequence_index = self.movement_counter_like_cpp;
-            self.movement_counter_like_cpp = self.movement_counter_like_cpp.wrapping_add(1);
-            return Some(sequence_index);
-        }
-        canonical
-    }
-
-    /// C++ `Player::SendInitialPacketsBeforeAddToMap` resets `m_movementCounter` to 0 for a
-    /// non-seamless add (login / far teleport). Player.cpp:23483.
-    pub(crate) fn reset_movement_counter_like_cpp(&mut self) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.unit_mut().reset_movement_counter_like_cpp();
-            })
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.movement_counter_like_cpp = 0;
-        }
-        canonical || cfg!(test) && self.player_handle_like_cpp.is_none()
-    }
-
-    /// Current `Unit::m_movementCounter` value (read without advancing). C++ reads this for
-    /// `SMSG_RESUME_TOKEN.SequenceIndex` on far teleport (MovementHandler.cpp:109), before
-    /// `SendInitialPacketsBeforeAddToMap` resets it.
-    pub(crate) fn movement_counter_like_cpp(&self) -> Option<u32> {
-        let canonical =
-            self.with_owned_player_like_cpp(|player| player.unit().movement_counter_like_cpp());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.movement_counter_like_cpp);
-        }
-        canonical
-    }
-
     /// C++ `CollectionMgr::SaveAccountMounts`.
     pub(crate) fn account_mount_save_rows_like_cpp(
         &self,
@@ -29566,15 +28427,6 @@ impl WorldSession {
         self.player_name.as_deref()
     }
 
-    pub(crate) fn player_position_like_cpp(&self) -> Option<wow_core::Position> {
-        let canonical = self.with_owned_player_like_cpp(|player| player.unit().world().position());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return self.player_position;
-        }
-        canonical
-    }
-
     pub(crate) fn player_map_id_like_cpp(&self) -> u16 {
         self.current_map_id
     }
@@ -29590,22 +28442,6 @@ impl WorldSession {
             return self.player_faction_template_like_cpp;
         }
         canonical.flatten()
-    }
-
-    fn resolved_player_movement_flags_like_cpp(&self) -> Option<MovementFlag> {
-        let canonical =
-            self.with_owned_player_like_cpp(|player| player.unit().movement_flags_like_cpp());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.player_movement_flags_like_cpp);
-        }
-        canonical
-    }
-
-    #[cfg(test)]
-    pub(crate) fn player_movement_flags_like_cpp(&self) -> MovementFlag {
-        self.resolved_player_movement_flags_like_cpp()
-            .expect("test Player movement owner must resolve")
     }
 
     #[cfg(test)]
@@ -29624,20 +28460,6 @@ impl WorldSession {
         #[cfg(test)]
         if canonical.is_none() && self.player_handle_like_cpp.is_none() {
             return Some(self.represented_can_swim_to_fly_transition_like_cpp);
-        }
-        canonical
-    }
-
-    fn resolved_mover_fixed_position_vehicle_like_cpp(&self) -> Option<bool> {
-        let canonical = self.with_owned_player_like_cpp(|player| {
-            player
-                .gameplay_state()
-                .movement_control
-                .mover_fixed_position_vehicle
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.represented_mover_fixed_position_vehicle_like_cpp);
         }
         canonical
     }
@@ -30645,16 +29467,6 @@ impl WorldSession {
             self.represented_battleground_leave_requests_like_cpp = self
                 .represented_battleground_leave_requests_like_cpp
                 .saturating_add(1);
-        }
-    }
-
-    fn maybe_leave_represented_battleground_on_far_teleport_like_cpp(&mut self, new_map: u32) {
-        if self
-            .player_battleground_state_snapshot_like_cpp()
-            .and_then(|state| state.represented_map_id)
-            .is_some_and(|bg_map_id| bg_map_id != new_map)
-        {
-            self.request_represented_battleground_leave_like_cpp();
         }
     }
 
@@ -31768,51 +30580,6 @@ impl WorldSession {
         self.apply_represented_resurrection_health_like_cpp(health);
     }
 
-    pub(crate) fn schedule_represented_resurrection_after_teleport_like_cpp(
-        &mut self,
-        request: PlayerResurrectionRequestLikeCpp,
-    ) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player
-                    .resurrection_state_mut_like_cpp()
-                    .delayed_after_teleport = Some(request);
-            })
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.represented_delayed_resurrection_after_teleport_like_cpp = Some(request);
-        }
-        canonical || cfg!(test) && self.player_handle_like_cpp.is_none()
-    }
-
-    pub(crate) fn process_represented_delayed_resurrection_after_teleport_like_cpp(&mut self) {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player
-                    .resurrection_state_mut_like_cpp()
-                    .delayed_after_teleport
-                    .take()
-            })
-            .flatten();
-        #[cfg(test)]
-        let request = canonical.or_else(|| {
-            (self.player_handle_like_cpp.is_none())
-                .then(|| {
-                    self.represented_delayed_resurrection_after_teleport_like_cpp
-                        .take()
-                })
-                .flatten()
-        });
-        #[cfg(not(test))]
-        let request = canonical;
-        let Some(request) = request else {
-            return;
-        };
-
-        self.apply_represented_resurrection_health_like_cpp(request.health);
-    }
-
     fn player_resurrection_state_snapshot_like_cpp(
         &self,
     ) -> Option<PlayerResurrectionStateLikeCpp> {
@@ -31838,142 +30605,6 @@ impl WorldSession {
     ) -> Option<PlayerResurrectionRequestLikeCpp> {
         self.player_resurrection_state_snapshot_like_cpp()
             .and_then(|state| state.request)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_delayed_resurrection_after_teleport_like_cpp(
-        &self,
-    ) -> Option<PlayerResurrectionRequestLikeCpp> {
-        self.player_resurrection_state_snapshot_like_cpp()
-            .and_then(|state| state.delayed_after_teleport)
-    }
-
-    fn resolved_fall_information_like_cpp(&self) -> Option<(u32, f32)> {
-        let canonical =
-            self.with_owned_player_like_cpp(|player| player.fall_information_like_cpp());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some((self.last_fall_time_like_cpp, self.last_fall_z_like_cpp));
-        }
-        canonical
-    }
-
-    pub(crate) fn set_fall_information_like_cpp(&mut self, time: u32, z: f32) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.set_fall_information_like_cpp(time, z);
-            })
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.last_fall_time_like_cpp = time;
-            self.last_fall_z_like_cpp = z;
-        }
-        canonical || cfg!(test) && self.player_handle_like_cpp.is_none()
-    }
-
-    pub(crate) fn update_fall_information_if_needed_like_cpp(
-        &mut self,
-        movement_info: &wow_packet::packets::movement::MovementInfo,
-        is_fall_land: bool,
-    ) {
-        let Some((last_fall_time, last_fall_z)) = self.resolved_fall_information_like_cpp() else {
-            return;
-        };
-        if last_fall_time >= movement_info.jump.fall_time
-            || last_fall_z <= movement_info.position.z
-            || is_fall_land
-        {
-            self.set_fall_information_like_cpp(
-                movement_info.jump.fall_time,
-                movement_info.position.z,
-            );
-        }
-    }
-
-    pub(crate) fn handle_fall_like_cpp(
-        &mut self,
-        movement_info: &wow_packet::packets::movement::MovementInfo,
-    ) -> Option<MovementFallDamageEvent> {
-        let (_, last_fall_z) = self.resolved_fall_information_like_cpp()?;
-        let damage_control = self.resolved_player_damage_control_like_cpp()?;
-        let z_diff = last_fall_z - movement_info.position.z;
-        let (_, max_health, player_is_alive) = self.resolved_player_vitals_like_cpp()?;
-        if z_diff < 14.57
-            || !player_is_alive
-            || self.player_is_game_master_like_cpp() == Some(true)
-            || self.resolved_has_represented_aura_effect_like_cpp(
-                RepresentedAuraEffectLikeCpp::Hover,
-            )?
-            || self.resolved_has_represented_aura_effect_like_cpp(
-                RepresentedAuraEffectLikeCpp::FeatherFall,
-            )?
-            || self
-                .resolved_has_represented_aura_effect_like_cpp(RepresentedAuraEffectLikeCpp::Fly)?
-            || damage_control.normal_damage_immune
-        {
-            return None;
-        }
-
-        let safe_fall = self.resolved_total_represented_aura_modifier_like_cpp(
-            RepresentedAuraEffectLikeCpp::SafeFall,
-        )?;
-        let damage_percent = 0.018 * (z_diff - safe_fall as f32) - 0.2426;
-        if damage_percent <= 0.0 {
-            return None;
-        }
-
-        let mut damage = (damage_percent * max_health as f32) as u32;
-        if damage_control.cheat_god {
-            damage = 0;
-        }
-        damage = (damage as f32
-            * self.resolved_total_represented_aura_multiplier_like_cpp(
-                RepresentedAuraEffectLikeCpp::ModifyFallDamagePct,
-            )?) as u32;
-        if self.player_has_visible_aura_spell_like_cpp(43_621)? {
-            damage = max_health / 2;
-        }
-        damage = damage.min(max_health);
-        if damage == 0 {
-            return None;
-        }
-
-        let requested_damage = if damage_control.environmental_damage_immune {
-            0
-        } else {
-            damage
-        };
-        let (_original_health, health_after, _, final_damage, killed_player) = self
-            .apply_owned_player_damage_like_cpp(
-                requested_damage,
-                wow_constants::DeathState::JustDied,
-            )?;
-        self.sync_player_registry_state_like_cpp();
-        if final_damage > 0
-            && let Some(player_guid) = self.player_guid()
-        {
-            self.send_player_health_update_like_cpp(player_guid, u64::from(health_after));
-            self.send_environmental_damage_log_like_cpp(
-                player_guid,
-                DAMAGE_FALL_LIKE_CPP,
-                damage,
-                0,
-                0,
-            );
-            if killed_player {
-                self.send_player_health_values_update_like_cpp(player_guid, 0);
-            }
-        }
-
-        let event = MovementFallDamageEvent {
-            z_diff,
-            damage,
-            final_damage,
-        };
-        #[cfg(test)]
-        self.fall_damage_events_like_cpp.push(event);
-        Some(event)
     }
 
     fn resolved_represented_total_stat_multiplier_for_stat_like_cpp(
@@ -32070,11 +30701,6 @@ impl WorldSession {
     pub(crate) fn represented_total_stat_buff_multipliers_like_cpp(&self) -> [f32; 5] {
         self.resolved_represented_total_stat_buff_multipliers_like_cpp()
             .expect("test Player aura owner must resolve")
-    }
-
-    #[cfg(test)]
-    pub(crate) fn fall_damage_events_like_cpp(&self) -> &[MovementFallDamageEvent] {
-        &self.fall_damage_events_like_cpp
     }
 
     pub(crate) fn player_min_height_like_cpp(&self, position: wow_core::Position) -> f32 {
@@ -32763,20 +31389,6 @@ impl WorldSession {
         &self,
     ) -> &[RepresentedCalendarAddEventLikeCpp] {
         &self.represented_calendar_add_events_like_cpp
-    }
-
-    #[cfg_attr(not(test), allow(unused_variables))]
-    pub(crate) fn calendar_remove_event_like_cpp(&mut self, event_id: u64) {
-        #[cfg(test)]
-        self.represented_calendar_remove_events_like_cpp
-            .push(RepresentedCalendarRemoveEventLikeCpp { event_id });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_calendar_remove_events_like_cpp(
-        &self,
-    ) -> &[RepresentedCalendarRemoveEventLikeCpp] {
-        &self.represented_calendar_remove_events_like_cpp
     }
 
     #[cfg(test)]
@@ -33776,39 +32388,6 @@ impl WorldSession {
         );
     }
 
-    pub(crate) fn player_moved_unit_guid_like_cpp(&self) -> Option<ObjectGuid> {
-        let canonical = self.with_owned_player_like_cpp(|player| {
-            player.unit().subsystems().control.unit_moved_by_me
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            let guid = if self.player_moved_unit_guid_like_cpp.is_empty() {
-                self.player_guid()?
-            } else {
-                self.player_moved_unit_guid_like_cpp
-            };
-            return Some(guid);
-        }
-        canonical.flatten()
-    }
-
-    pub fn set_player_moved_unit_guid_like_cpp(&mut self, guid: ObjectGuid) {
-        #[cfg_attr(not(test), allow(unused_variables))]
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player
-                    .unit_mut()
-                    .subsystems_mut()
-                    .control
-                    .set_moved_unit((!guid.is_empty()).then_some(guid));
-            })
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.player_moved_unit_guid_like_cpp = guid;
-        }
-    }
-
     pub(crate) fn represented_request_vehicle_exit_like_cpp(&mut self) -> bool {
         let Some(seat_flags) = self
             .player_vehicle_seat_state_like_cpp()
@@ -33877,20 +32456,6 @@ impl WorldSession {
         &self.represented_vehicle_enter_requests_like_cpp
     }
 
-    #[cfg(test)]
-    pub(crate) fn represented_vehicle_dismiss_movements_like_cpp(
-        &self,
-    ) -> &[RepresentedVehicleDismissMovementLikeCpp] {
-        &self.represented_vehicle_dismiss_movements_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_vehicle_base_movements_like_cpp(
-        &self,
-    ) -> &[RepresentedVehicleBaseMovementLikeCpp] {
-        &self.represented_vehicle_base_movements_like_cpp
-    }
-
     fn represented_vehicle_base_guid_for_switch_like_cpp(&self) -> Option<ObjectGuid> {
         if self
             .player_vehicle_seat_state_like_cpp()
@@ -33946,54 +32511,6 @@ impl WorldSession {
             }
             _ => false,
         }
-    }
-
-    pub(crate) fn represented_move_change_vehicle_seats_like_cpp(
-        &mut self,
-        status: &mut wow_packet::packets::movement::MovementInfo,
-        dst_vehicle: ObjectGuid,
-        dst_seat_index: u8,
-    ) -> bool {
-        let Some(vehicle_base_guid) = self.represented_vehicle_base_guid_for_switch_like_cpp()
-        else {
-            return false;
-        };
-        if !self.represented_current_vehicle_seat_can_switch_from_like_cpp() {
-            return false;
-        }
-
-        self.sanitize_movement_info_represented_like_cpp(status);
-        if status.guid != vehicle_base_guid {
-            return false;
-        }
-
-        let dst_seat_id = dst_seat_index as i8;
-        let dst_vehicle_exists_with_empty_seat = !dst_vehicle.is_empty()
-            && self.represented_vehicle_seat_spell_click_plan_available_like_cpp(
-                dst_vehicle,
-                dst_seat_id,
-            );
-        let action = crate::handlers::vehicle::move_change_vehicle_seats_action_like_cpp(
-            true,
-            true,
-            vehicle_base_guid,
-            status.guid,
-            dst_vehicle,
-            dst_seat_index,
-            dst_vehicle_exists_with_empty_seat,
-        );
-
-        #[cfg(test)]
-        self.represented_vehicle_base_movements_like_cpp.push(
-            RepresentedVehicleBaseMovementLikeCpp {
-                vehicle_guid: vehicle_base_guid,
-                sanitized_flags: status.flags,
-                position: status.position,
-                time: status.time,
-            },
-        );
-        let _ = self.record_represented_vehicle_seat_action_like_cpp(action);
-        true
     }
 
     pub(crate) fn represented_request_vehicle_switch_seat_like_cpp(
@@ -36016,12 +34533,6 @@ impl WorldSession {
         );
     }
 
-    fn send_represented_capture_point_removed_like_cpp(&mut self, gameobject_guid: ObjectGuid) {
-        self.send_packet(&wow_packet::packets::misc::CapturePointRemoved {
-            capture_point_guid: gameobject_guid,
-        });
-    }
-
     pub(crate) fn apply_represented_new_flag_state_command_like_cpp(
         &mut self,
         gameobject_guid: ObjectGuid,
@@ -37119,14 +35630,6 @@ impl WorldSession {
         self.sync_current_player_session_visibility_detection_like_cpp();
     }
 
-    fn clear_active_player_transport_server_time_override_for_far_teleport_like_cpp(&mut self) {
-        let _ = self.mutate_active_player_update_state_like_cpp(|state| {
-            state.active_local_flags &= !PLAYER_LOCAL_FLAG_OVERRIDE_TRANSPORT_SERVER_TIME_LIKE_CPP;
-            state.active_transport_server_time = 0;
-        });
-        self.sync_current_player_session_visibility_detection_like_cpp();
-    }
-
     #[cfg(test)]
     pub(crate) fn active_player_transport_server_time_like_cpp(&self) -> i32 {
         self.active_player_update_state_like_cpp()
@@ -37563,280 +36066,6 @@ impl WorldSession {
         true
     }
 
-    pub(crate) fn validate_movement_ack_status_like_cpp(
-        &self,
-        status: &wow_packet::packets::movement::MovementInfo,
-    ) -> bool {
-        let Some(player_guid) = self.player_guid() else {
-            return false;
-        };
-
-        status.guid == player_guid && status.position.is_valid_map_coord_like_cpp()
-    }
-
-    pub(crate) fn validate_and_sanitize_movement_ack_status_represented_like_cpp(
-        &self,
-        status: &mut wow_packet::packets::movement::MovementInfo,
-    ) -> bool {
-        if !self.validate_movement_ack_status_like_cpp(status) {
-            return false;
-        }
-
-        self.sanitize_movement_info_flags_represented_like_cpp(status);
-        true
-    }
-
-    pub(crate) fn sanitize_movement_info_flags_represented_like_cpp(
-        &self,
-        movement_info: &mut wow_packet::packets::movement::MovementInfo,
-    ) -> MovementFlag {
-        self.sanitize_movement_info_represented_like_cpp(movement_info)
-            .removed_flags
-    }
-
-    pub(crate) fn sanitize_movement_info_represented_like_cpp(
-        &self,
-        movement_info: &mut wow_packet::packets::movement::MovementInfo,
-    ) -> wow_anticheat::ValidationResult {
-        wow_anticheat::validate_movement_info(
-            movement_info,
-            &wow_anticheat::PlayerState {
-                // Fixed-position is an authorization proof. An unresolved
-                // Player must not preserve a client-supplied ROOT flag.
-                mover_fixed_position_vehicle: self.resolved_mover_fixed_position_vehicle_like_cpp()
-                    == Some(true),
-                // An unresolved/stale Player may not authorize client-only
-                // movement flags. Treat the proof as absent for sanitization,
-                // without claiming that the canonical aura set is empty.
-                has_hover_aura: self.resolved_has_represented_aura_effect_like_cpp(
-                    RepresentedAuraEffectLikeCpp::Hover,
-                ) == Some(true),
-                has_water_walk_aura: self.resolved_has_represented_aura_effect_like_cpp(
-                    RepresentedAuraEffectLikeCpp::WaterWalk,
-                ) == Some(true),
-                has_ghost_aura: self.resolved_has_represented_aura_effect_like_cpp(
-                    RepresentedAuraEffectLikeCpp::Ghost,
-                ) == Some(true),
-                has_feather_fall_aura: self.resolved_has_represented_aura_effect_like_cpp(
-                    RepresentedAuraEffectLikeCpp::FeatherFall,
-                ) == Some(true),
-                has_fly_aura: self.resolved_has_represented_aura_effect_like_cpp(
-                    RepresentedAuraEffectLikeCpp::Fly,
-                ) == Some(true),
-                has_mounted_flight_speed_aura: self.resolved_has_represented_aura_effect_like_cpp(
-                    RepresentedAuraEffectLikeCpp::MountedFlightSpeed,
-                ) == Some(true),
-                is_player_security: self.player_is_game_master_like_cpp() != Some(true),
-            },
-        )
-    }
-
-    pub(crate) fn record_movement_ack_event_like_cpp(&mut self, event: MovementAckEventLikeCpp) {
-        #[cfg(test)]
-        self.movement_ack_events_like_cpp.push(event);
-        #[cfg(not(test))]
-        let _ = event;
-    }
-
-    pub(crate) fn apply_knock_back_ack_like_cpp(
-        &mut self,
-        opcode: ClientOpcodes,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-    ) -> bool {
-        if !self.validate_and_sanitize_movement_ack_status_represented_like_cpp(&mut ack.status) {
-            self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-                opcode,
-                mover_guid: ack.status.guid,
-                ack_index: Some(ack.ack_index),
-                movement_force_id: None,
-                movement_force_type: None,
-                adjusted_time: None,
-                speed: None,
-                time_skipped: None,
-                spline_id: None,
-                accepted: false,
-            });
-            return false;
-        }
-
-        let mut status = ack.status.clone();
-        status.time = self.adjust_client_movement_time_like_cpp(status.time);
-        self.set_player_movement_time_like_cpp(status.time);
-        self.set_player_movement_flags_like_cpp(status.flags);
-        self.set_player_position_like_cpp(status.position);
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode,
-            mover_guid: status.guid,
-            ack_index: Some(ack.ack_index),
-            movement_force_id: None,
-            movement_force_type: None,
-            adjusted_time: Some(status.time),
-            speed: None,
-            time_skipped: None,
-            spline_id: None,
-            accepted: true,
-        });
-        true
-    }
-
-    pub(crate) fn apply_move_time_skipped_like_cpp(
-        &mut self,
-        mover_guid: ObjectGuid,
-        time_skipped: u32,
-    ) -> bool {
-        let adjusted_time = (self.player_guid() == Some(mover_guid))
-            .then(|| {
-                self.resolved_player_movement_time_like_cpp()
-                    .map(|time| time.saturating_add(time_skipped))
-            })
-            .flatten();
-        if let Some(adjusted_time) = adjusted_time {
-            self.set_player_movement_time_like_cpp(adjusted_time);
-        }
-        let accepted = adjusted_time.is_some();
-
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode: ClientOpcodes::MoveTimeSkipped,
-            mover_guid,
-            ack_index: None,
-            movement_force_id: None,
-            movement_force_type: None,
-            adjusted_time,
-            speed: None,
-            time_skipped: Some(time_skipped),
-            spline_id: None,
-            accepted,
-        });
-        accepted
-    }
-
-    pub(crate) fn record_validated_movement_ack_like_cpp(
-        &mut self,
-        opcode: ClientOpcodes,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-        speed: Option<f32>,
-    ) -> bool {
-        let accepted =
-            self.validate_and_sanitize_movement_ack_status_represented_like_cpp(&mut ack.status);
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode,
-            mover_guid: ack.status.guid,
-            ack_index: Some(ack.ack_index),
-            movement_force_id: None,
-            movement_force_type: None,
-            adjusted_time: None,
-            speed,
-            time_skipped: None,
-            spline_id: None,
-            accepted,
-        });
-        accepted
-    }
-
-    pub(crate) fn apply_move_set_vehicle_rec_id_ack_like_cpp(
-        &mut self,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-    ) {
-        self.sanitize_movement_info_represented_like_cpp(&mut ack.status);
-    }
-
-    pub(crate) fn record_apply_movement_force_ack_like_cpp(
-        &mut self,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-        force: &wow_packet::packets::movement::MovementForce,
-    ) -> bool {
-        if !self.validate_and_sanitize_movement_ack_status_represented_like_cpp(&mut ack.status) {
-            self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-                opcode: ClientOpcodes::MoveApplyMovementForceAck,
-                mover_guid: ack.status.guid,
-                ack_index: Some(ack.ack_index),
-                movement_force_id: Some(force.id),
-                movement_force_type: Some(force.force_type.to_wire()),
-                adjusted_time: None,
-                speed: None,
-                time_skipped: None,
-                spline_id: None,
-                accepted: false,
-            });
-            return false;
-        }
-
-        let adjusted_time = self.adjust_client_movement_time_like_cpp(ack.status.time);
-        ack.status.time = adjusted_time;
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode: ClientOpcodes::MoveApplyMovementForceAck,
-            mover_guid: ack.status.guid,
-            ack_index: Some(ack.ack_index),
-            movement_force_id: Some(force.id),
-            movement_force_type: Some(force.force_type.to_wire()),
-            adjusted_time: Some(adjusted_time),
-            speed: None,
-            time_skipped: None,
-            spline_id: None,
-            accepted: true,
-        });
-        true
-    }
-
-    pub(crate) fn record_remove_movement_force_ack_like_cpp(
-        &mut self,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-        force_id: ObjectGuid,
-    ) -> bool {
-        if !self.validate_and_sanitize_movement_ack_status_represented_like_cpp(&mut ack.status) {
-            self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-                opcode: ClientOpcodes::MoveRemoveMovementForceAck,
-                mover_guid: ack.status.guid,
-                ack_index: Some(ack.ack_index),
-                movement_force_id: Some(force_id),
-                movement_force_type: None,
-                adjusted_time: None,
-                speed: None,
-                time_skipped: None,
-                spline_id: None,
-                accepted: false,
-            });
-            return false;
-        }
-
-        let adjusted_time = self.adjust_client_movement_time_like_cpp(ack.status.time);
-        ack.status.time = adjusted_time;
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode: ClientOpcodes::MoveRemoveMovementForceAck,
-            mover_guid: ack.status.guid,
-            ack_index: Some(ack.ack_index),
-            movement_force_id: Some(force_id),
-            movement_force_type: None,
-            adjusted_time: Some(adjusted_time),
-            speed: None,
-            time_skipped: None,
-            spline_id: None,
-            accepted: true,
-        });
-        true
-    }
-
-    pub(crate) fn record_move_spline_done_like_cpp(
-        &mut self,
-        status: &mut wow_packet::packets::movement::MovementInfo,
-        spline_id: i32,
-    ) -> bool {
-        let accepted = self.validate_and_sanitize_movement_ack_status_represented_like_cpp(status);
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode: ClientOpcodes::MoveSplineDone,
-            mover_guid: status.guid,
-            ack_index: None,
-            movement_force_id: None,
-            movement_force_type: None,
-            adjusted_time: None,
-            speed: None,
-            time_skipped: None,
-            spline_id: Some(spline_id),
-            accepted,
-        });
-        accepted
-    }
-
     pub(crate) fn player_taxi_state_snapshot_like_cpp(
         &self,
     ) -> Option<wow_entities::PlayerTaxiState> {
@@ -37890,478 +36119,6 @@ impl WorldSession {
         }
         // PlayerTaxi mutates the owning Player's route, not a Session copy.
         self.with_owned_player_mut_like_cpp(|player| f(&mut player.gameplay_state_mut().taxi))
-    }
-
-    pub(crate) fn handle_move_spline_done_taxi_like_cpp(
-        &mut self,
-        status: &mut wow_packet::packets::movement::MovementInfo,
-        spline_id: i32,
-    ) -> MoveSplineDoneTaxiActionLikeCpp {
-        if !self.record_move_spline_done_like_cpp(status, spline_id) {
-            return self.record_move_spline_done_taxi_event_like_cpp(
-                spline_id,
-                MoveSplineDoneTaxiActionLikeCpp::InvalidMovement,
-                None,
-                None,
-                None,
-                false,
-            );
-        }
-
-        let Some(taxi_state) = self.player_taxi_state_snapshot_like_cpp() else {
-            return self.record_move_spline_done_taxi_event_like_cpp(
-                spline_id,
-                MoveSplineDoneTaxiActionLikeCpp::IgnoredUnexpectedFinalPath,
-                None,
-                None,
-                None,
-                false,
-            );
-        };
-        let current_destination = taxi_state.destinations.get(1).copied();
-        if let Some(destination_node_id) = current_destination {
-            let Some(flight) = taxi_state.flight else {
-                return self.record_move_spline_done_taxi_event_like_cpp(
-                    spline_id,
-                    MoveSplineDoneTaxiActionLikeCpp::InProgressNoFlightGenerator,
-                    Some(destination_node_id),
-                    None,
-                    None,
-                    false,
-                );
-            };
-
-            let destination_map_id = self
-                .taxi_node_map_ids_like_cpp
-                .get(&destination_node_id)
-                .copied();
-            let should_teleport = destination_map_id
-                .map(|map_id| map_id != self.player_map_id_like_cpp())
-                .unwrap_or(false)
-                || flight.current_node.teleport_flag;
-
-            if should_teleport {
-                if let (Some(map_id), Some(node)) = (destination_map_id, flight.node_after_teleport)
-                {
-                    if self
-                        .mutate_player_taxi_state_like_cpp(|taxi| {
-                            taxi.flight = Some(wow_entities::PlayerTaxiFlightStateLikeCpp {
-                                current_node: node,
-                                node_after_teleport: None,
-                            });
-                        })
-                        .is_none()
-                    {
-                        return self.record_move_spline_done_taxi_event_like_cpp(
-                            spline_id,
-                            MoveSplineDoneTaxiActionLikeCpp::IgnoredUnexpectedFinalPath,
-                            Some(destination_node_id),
-                            None,
-                            None,
-                            false,
-                        );
-                    }
-                    self.set_player_map_position_like_cpp(map_id, node.position);
-                    return self.record_move_spline_done_taxi_event_like_cpp(
-                        spline_id,
-                        MoveSplineDoneTaxiActionLikeCpp::TeleportRequested,
-                        Some(destination_node_id),
-                        Some(map_id),
-                        Some(node.position),
-                        false,
-                    );
-                }
-            }
-
-            return self.record_move_spline_done_taxi_event_like_cpp(
-                spline_id,
-                MoveSplineDoneTaxiActionLikeCpp::InProgressNoTeleport,
-                Some(destination_node_id),
-                None,
-                None,
-                false,
-            );
-        }
-
-        if taxi_state.destinations.len() != 1 {
-            return self.record_move_spline_done_taxi_event_like_cpp(
-                spline_id,
-                MoveSplineDoneTaxiActionLikeCpp::IgnoredUnexpectedFinalPath,
-                None,
-                None,
-                None,
-                false,
-            );
-        }
-
-        if self
-            .mutate_player_taxi_state_like_cpp(|taxi| {
-                taxi.destinations.clear();
-                taxi.flight = None;
-                taxi.mounted = false;
-                taxi.unit_flags &= !(UnitFlags::REMOVE_CLIENT_CONTROL | UnitFlags::ON_TAXI).bits();
-            })
-            .is_none()
-        {
-            return self.record_move_spline_done_taxi_event_like_cpp(
-                spline_id,
-                MoveSplineDoneTaxiActionLikeCpp::IgnoredUnexpectedFinalPath,
-                None,
-                None,
-                None,
-                false,
-            );
-        }
-        let current_z = self
-            .player_position_like_cpp()
-            .map(|position| position.z)
-            .unwrap_or(status.position.z);
-        self.set_fall_information_like_cpp(0, current_z);
-        let honorless_target_cast = self
-            .player_world_local_state_like_cpp()
-            .is_some_and(|state| state.pvp_hostile);
-
-        self.record_move_spline_done_taxi_event_like_cpp(
-            spline_id,
-            MoveSplineDoneTaxiActionLikeCpp::FinalCleanup,
-            None,
-            None,
-            None,
-            honorless_target_cast,
-        )
-    }
-
-    fn record_move_spline_done_taxi_event_like_cpp(
-        &mut self,
-        spline_id: i32,
-        action: MoveSplineDoneTaxiActionLikeCpp,
-        destination_node_id: Option<u32>,
-        teleport_map_id: Option<u16>,
-        teleport_position: Option<wow_core::Position>,
-        honorless_target_cast: bool,
-    ) -> MoveSplineDoneTaxiActionLikeCpp {
-        #[cfg(test)]
-        self.move_spline_done_taxi_events_like_cpp
-            .push(MoveSplineDoneTaxiEventLikeCpp {
-                spline_id,
-                action,
-                destination_node_id,
-                teleport_map_id,
-                teleport_position,
-                honorless_target_cast,
-            });
-        #[cfg(not(test))]
-        let _ = (
-            spline_id,
-            destination_node_id,
-            teleport_map_id,
-            teleport_position,
-            honorless_target_cast,
-        );
-        action
-    }
-
-    pub(crate) fn record_move_teleport_ack_like_cpp(
-        &mut self,
-        mover_guid: ObjectGuid,
-        ack_index: i32,
-        move_time: i32,
-    ) -> bool {
-        let accepted = self.player_guid() == Some(mover_guid);
-        self.record_movement_ack_event_like_cpp(MovementAckEventLikeCpp {
-            opcode: ClientOpcodes::MoveTeleportAck,
-            mover_guid,
-            ack_index: Some(ack_index),
-            movement_force_id: None,
-            movement_force_type: None,
-            adjusted_time: (move_time >= 0).then_some(move_time as u32),
-            speed: None,
-            time_skipped: None,
-            spline_id: None,
-            accepted,
-        });
-        accepted
-    }
-
-    pub(crate) fn handle_move_teleport_ack_like_cpp(
-        &mut self,
-        mover_guid: ObjectGuid,
-        ack_index: i32,
-        move_time: i32,
-    ) -> MoveTeleportAckActionLikeCpp {
-        let accepted = self.record_move_teleport_ack_like_cpp(mover_guid, ack_index, move_time);
-        let Some(teleport) = self.player_teleport_state_snapshot_like_cpp() else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-        if !teleport.near_pending {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::NotBeingTeleportedNear,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        }
-
-        if !accepted {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::WrongMover,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        }
-
-        let Some((map_id, destination)) = teleport.near_destination else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingDestination,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-
-        let Some(world_local_before) = self.player_world_local_state_like_cpp() else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-        let Some(player_guid) = self.player_guid() else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-        let Some(pvp_enabled_before) = self.player_is_pvp_like_cpp(player_guid) else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-        let Some(in_pvp_before) = self.player_has_in_pvp_flag_like_cpp(player_guid) else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                None,
-                None,
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-
-        if !self.update_player_teleport_state_like_cpp(|state| state.near_pending = false) {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                Some(map_id),
-                Some(destination),
-                None,
-                None,
-                None,
-                false,
-                false,
-                false,
-                false,
-            );
-        }
-        let old_zone = world_local_before.zone_id;
-        self.set_player_map_position_like_cpp(map_id, destination);
-        self.update_registry_position();
-        self.set_fall_information_like_cpp(0, destination.z);
-
-        let (new_zone, new_area) = teleport
-            .near_destination_zone_area
-            .unwrap_or((world_local_before.zone_id, world_local_before.area_id));
-        self.update_zone_represented_like_cpp(new_zone, new_area);
-
-        let zone_changed = old_zone != new_zone;
-        let Some(world_local_after) = self.player_world_local_state_like_cpp() else {
-            return self.record_move_teleport_ack_event_like_cpp(
-                mover_guid,
-                ack_index,
-                move_time,
-                MoveTeleportAckActionLikeCpp::MissingPlayerOwner,
-                Some(map_id),
-                Some(destination),
-                Some(old_zone),
-                Some(new_zone),
-                Some(new_area),
-                false,
-                false,
-                false,
-                false,
-            );
-        };
-        let honorless_target_cast = zone_changed && world_local_after.pvp_hostile;
-        let pvp_disabled =
-            zone_changed && !world_local_after.pvp_hostile && pvp_enabled_before && !in_pvp_before;
-        if pvp_disabled {
-            self.update_player_pvp_like_cpp(false, true);
-        }
-
-        self.resummon_pet_temporary_unsummoned_like_cpp();
-        self.process_represented_delayed_resurrection_after_teleport_like_cpp();
-        #[cfg(test)]
-        {
-            self.delayed_operations_processed_like_cpp =
-                self.delayed_operations_processed_like_cpp.saturating_add(1);
-        }
-
-        self.record_move_teleport_ack_event_like_cpp(
-            mover_guid,
-            ack_index,
-            move_time,
-            MoveTeleportAckActionLikeCpp::Accepted,
-            Some(map_id),
-            Some(destination),
-            Some(old_zone),
-            Some(new_zone),
-            Some(new_area),
-            honorless_target_cast,
-            pvp_disabled,
-            true,
-            true,
-        )
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn record_move_teleport_ack_event_like_cpp(
-        &mut self,
-        mover_guid: ObjectGuid,
-        ack_index: i32,
-        move_time: i32,
-        action: MoveTeleportAckActionLikeCpp,
-        destination_map_id: Option<u16>,
-        destination_position: Option<wow_core::Position>,
-        old_zone_id: Option<u32>,
-        new_zone_id: Option<u32>,
-        new_area_id: Option<u32>,
-        honorless_target_cast: bool,
-        pvp_disabled: bool,
-        pet_resummon_requested: bool,
-        delayed_operations_processed: bool,
-    ) -> MoveTeleportAckActionLikeCpp {
-        #[cfg(test)]
-        self.move_teleport_ack_events_like_cpp
-            .push(MoveTeleportAckEventLikeCpp {
-                mover_guid,
-                ack_index,
-                move_time,
-                action,
-                destination_map_id,
-                destination_position,
-                old_zone_id,
-                new_zone_id,
-                new_area_id,
-                honorless_target_cast,
-                pvp_disabled,
-                pet_resummon_requested,
-                delayed_operations_processed,
-            });
-        #[cfg(not(test))]
-        let _ = (
-            mover_guid,
-            ack_index,
-            move_time,
-            destination_map_id,
-            destination_position,
-            old_zone_id,
-            new_zone_id,
-            new_area_id,
-            honorless_target_cast,
-            pvp_disabled,
-            pet_resummon_requested,
-            delayed_operations_processed,
-        );
-        action
-    }
-
-    #[cfg(test)]
-    pub(crate) fn movement_ack_events_like_cpp(&self) -> &[MovementAckEventLikeCpp] {
-        &self.movement_ack_events_like_cpp
     }
 
     #[cfg(test)]
@@ -38466,34 +36223,6 @@ impl WorldSession {
             return Some(result);
         }
         canonical.then_some(result)
-    }
-
-    pub(crate) fn remove_represented_at_login_flag_like_cpp(
-        &mut self,
-        flags: u16,
-        persist: bool,
-    ) -> bool {
-        let Some(removed) = self.mutate_player_persistent_capability_state_like_cpp(|state| {
-            let removed = (state.at_login_flags & flags) != 0;
-            state.at_login_flags &= !flags;
-            removed
-        }) else {
-            return false;
-        };
-        if !removed {
-            return false;
-        }
-        if persist {
-            #[cfg(test)]
-            self.represented_at_login_flag_removals_like_cpp.push(
-                RepresentedAtLoginFlagRemovalLikeCpp {
-                    flags,
-                    persist,
-                    db_statement_unrepresented: true,
-                },
-            );
-        }
-        true
     }
 
     pub(crate) fn set_represented_at_login_flags_like_cpp(&mut self, flags: u16) -> bool {
@@ -38618,19 +36347,6 @@ impl WorldSession {
         });
     }
 
-    #[cfg(test)]
-    pub(crate) fn fall_information_like_cpp(&self) -> (u32, f32) {
-        self.resolved_fall_information_like_cpp()
-            .expect("test Player fall-information owner must resolve")
-    }
-
-    #[cfg(test)]
-    pub(crate) fn move_spline_done_taxi_events_like_cpp(
-        &self,
-    ) -> &[MoveSplineDoneTaxiEventLikeCpp] {
-        &self.move_spline_done_taxi_events_like_cpp
-    }
-
     pub(crate) fn set_player_zone_area_like_cpp(&mut self, zone_id: u32, area_id: u32) {
         let changed = self
             .player_zone_area_like_cpp()
@@ -38693,116 +36409,6 @@ impl WorldSession {
     }
 
     #[cfg(test)]
-    pub(crate) fn set_near_teleport_pending_like_cpp(
-        &mut self,
-        pending: bool,
-        destination: Option<(u16, wow_core::Position)>,
-        zone_area: Option<(u32, u32)>,
-    ) -> bool {
-        self.update_player_teleport_state_like_cpp(|state| {
-            state.near_pending = pending;
-            state.near_destination = destination;
-            state.near_destination_zone_area = zone_area;
-        })
-    }
-
-    fn player_teleport_state_snapshot_like_cpp(&self) -> Option<PlayerTeleportStateLikeCpp> {
-        let canonical = self.with_owned_player_like_cpp(|player| *player.teleport_state_like_cpp());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(PlayerTeleportStateLikeCpp {
-                recovery: Default::default(),
-                far_destination: self.pending_teleport,
-                post_add: None,
-                can_delay: self.represented_can_delay_teleport_like_cpp,
-                has_delayed: self.represented_has_delayed_teleport_like_cpp,
-                near_pending: self.near_teleport_pending_like_cpp,
-                far_pending: self.represented_far_teleport_pending_like_cpp,
-                near_destination: self.near_teleport_destination_like_cpp,
-                delayed: self.represented_delayed_teleport_like_cpp,
-                near_destination_zone_area: self.near_teleport_destination_zone_area_like_cpp,
-            });
-        }
-        canonical
-    }
-
-    fn update_player_teleport_state_like_cpp(
-        &mut self,
-        update: impl FnOnce(&mut PlayerTeleportStateLikeCpp),
-    ) -> bool {
-        if self.player_handle_like_cpp.is_some() {
-            return self
-                .with_owned_player_mut_like_cpp(|player| {
-                    update(player.teleport_state_mut_like_cpp())
-                })
-                .is_some();
-        }
-        #[cfg(test)]
-        {
-            let mut state = self
-                .player_teleport_state_snapshot_like_cpp()
-                .unwrap_or_default();
-            update(&mut state);
-            self.pending_teleport = state.far_destination;
-            self.represented_can_delay_teleport_like_cpp = state.can_delay;
-            self.represented_has_delayed_teleport_like_cpp = state.has_delayed;
-            self.near_teleport_pending_like_cpp = state.near_pending;
-            self.represented_far_teleport_pending_like_cpp = state.far_pending;
-            self.near_teleport_destination_like_cpp = state.near_destination;
-            self.represented_delayed_teleport_like_cpp = state.delayed;
-            self.near_teleport_destination_zone_area_like_cpp = state.near_destination_zone_area;
-            true
-        }
-        #[cfg(not(test))]
-        {
-            let _ = update;
-            false
-        }
-    }
-
-    pub(crate) fn set_represented_can_delay_teleport_like_cpp(&mut self, can_delay: bool) -> bool {
-        self.update_player_teleport_state_like_cpp(|state| state.can_delay = can_delay)
-    }
-
-    pub(crate) fn represented_can_delay_teleport_like_cpp(&self) -> bool {
-        self.player_teleport_state_snapshot_like_cpp()
-            .is_some_and(|state| state.can_delay)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_has_delayed_teleport_like_cpp(&self) -> bool {
-        self.player_teleport_state_snapshot_like_cpp()
-            .is_some_and(|state| state.has_delayed)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_delayed_teleport_like_cpp(
-        &self,
-    ) -> Option<(u32, wow_core::Position, TeleportToOptionsLikeCpp)> {
-        self.player_teleport_state_snapshot_like_cpp()
-            .and_then(|state| state.delayed)
-    }
-
-    pub(crate) fn represented_far_teleport_pending_like_cpp(&self) -> bool {
-        self.player_teleport_state_snapshot_like_cpp()
-            .is_some_and(|state| state.far_pending)
-    }
-
-    pub(crate) fn set_represented_far_teleport_pending_like_cpp(&mut self, pending: bool) -> bool {
-        self.update_player_teleport_state_like_cpp(|state| state.far_pending = pending)
-    }
-
-    pub(crate) fn near_teleport_pending_like_cpp(&self) -> bool {
-        self.player_teleport_state_snapshot_like_cpp()
-            .is_some_and(|state| state.near_pending)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn move_teleport_ack_events_like_cpp(&self) -> &[MoveTeleportAckEventLikeCpp] {
-        &self.move_teleport_ack_events_like_cpp
-    }
-
-    #[cfg(test)]
     pub(crate) fn temporary_pet_resummon_requests_like_cpp(&self) -> u32 {
         self.temporary_pet_resummon_requests_like_cpp
     }
@@ -38810,307 +36416,6 @@ impl WorldSession {
     #[cfg(test)]
     pub(crate) fn delayed_operations_processed_like_cpp(&self) -> u32 {
         self.delayed_operations_processed_like_cpp
-    }
-
-    pub(crate) fn resolved_player_movement_time_like_cpp(&self) -> Option<u32> {
-        let canonical =
-            self.with_owned_player_like_cpp(|player| player.unit().movement_time_like_cpp());
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.player_movement_time_like_cpp);
-        }
-        canonical
-    }
-
-    #[cfg(test)]
-    pub(crate) fn player_movement_time_like_cpp(&self) -> u32 {
-        self.resolved_player_movement_time_like_cpp()
-            .expect("test Player movement-time owner must resolve")
-    }
-
-    #[cfg(test)]
-    pub(crate) fn player_movement_jump_like_cpp(&self) -> &wow_packet::packets::movement::JumpInfo {
-        &self.player_movement_jump_like_cpp
-    }
-
-    pub(crate) fn movement_speed_ack_move_type_like_cpp(
-        opcode: ClientOpcodes,
-    ) -> Option<UnitMoveTypeLikeCpp> {
-        match opcode {
-            ClientOpcodes::MoveForceWalkSpeedChangeAck => Some(UnitMoveTypeLikeCpp::Walk),
-            ClientOpcodes::MoveForceRunSpeedChangeAck => Some(UnitMoveTypeLikeCpp::Run),
-            ClientOpcodes::MoveForceRunBackSpeedChangeAck => Some(UnitMoveTypeLikeCpp::RunBack),
-            ClientOpcodes::MoveForceSwimSpeedChangeAck => Some(UnitMoveTypeLikeCpp::Swim),
-            ClientOpcodes::MoveForceSwimBackSpeedChangeAck => Some(UnitMoveTypeLikeCpp::SwimBack),
-            ClientOpcodes::MoveForceTurnRateChangeAck => Some(UnitMoveTypeLikeCpp::TurnRate),
-            ClientOpcodes::MoveForceFlightSpeedChangeAck => Some(UnitMoveTypeLikeCpp::Flight),
-            ClientOpcodes::MoveForceFlightBackSpeedChangeAck => {
-                Some(UnitMoveTypeLikeCpp::FlightBack)
-            }
-            ClientOpcodes::MoveForcePitchRateChangeAck => Some(UnitMoveTypeLikeCpp::PitchRate),
-            _ => None,
-        }
-    }
-
-    fn resolved_player_movement_speed_rate_like_cpp(
-        &self,
-        move_type: UnitMoveTypeLikeCpp,
-    ) -> Option<f32> {
-        let index = move_type.index();
-        let canonical = self
-            .with_owned_player_like_cpp(|player| player.unit().speed_rate_at_like_cpp(index))
-            .flatten();
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.movement_speed_rates_like_cpp[index]);
-        }
-        canonical
-    }
-
-    fn set_player_movement_speed_rate_like_cpp_inner(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-        rate: f32,
-    ) -> bool {
-        let index = move_type.index();
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.unit_mut().set_speed_rate_at_like_cpp(index, rate)
-            })
-            .unwrap_or(false);
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.movement_speed_rates_like_cpp[index] = rate;
-        }
-        canonical || cfg!(test) && self.player_handle_like_cpp.is_none()
-    }
-
-    fn resolved_player_movement_speed_like_cpp(
-        &self,
-        move_type: UnitMoveTypeLikeCpp,
-    ) -> Option<f32> {
-        Some(
-            PLAYER_BASE_MOVE_SPEED_LIKE_CPP[move_type.index()]
-                * self.resolved_player_movement_speed_rate_like_cpp(move_type)?,
-        )
-    }
-
-    #[cfg(test)]
-    pub(crate) fn player_movement_speed_like_cpp(&self, move_type: UnitMoveTypeLikeCpp) -> f32 {
-        self.resolved_player_movement_speed_like_cpp(move_type)
-            .expect("test Player movement-speed owner must resolve")
-    }
-
-    fn represented_run_speed_rate_like_cpp(&self) -> Option<f32> {
-        let mounted = self.resolved_player_mounted_like_cpp()?;
-        let (main_mod_effect, stack_effect, not_stack_effect) = if mounted {
-            (
-                RepresentedAuraEffectLikeCpp::MountedSpeed,
-                RepresentedAuraEffectLikeCpp::MountedSpeedAlways,
-                RepresentedAuraEffectLikeCpp::MountedSpeedNotStack,
-            )
-        } else {
-            (
-                RepresentedAuraEffectLikeCpp::Speed,
-                RepresentedAuraEffectLikeCpp::SpeedAlways,
-                RepresentedAuraEffectLikeCpp::SpeedNotStack,
-            )
-        };
-        let main_mod = self.max_represented_aura_amount_like_cpp(main_mod_effect)?;
-        let stack_bonus = self.total_represented_aura_amount_multiplier_like_cpp(stack_effect)?;
-        let not_stack_bonus = 1.0
-            + self
-                .max_represented_aura_amount_like_cpp(not_stack_effect)?
-                .max(0) as f32
-                / 100.0;
-
-        let speed = stack_bonus.max(not_stack_bonus) * (1.0 + main_mod.max(0) as f32 / 100.0);
-        Some(self.apply_represented_forward_speed_adjustments_like_cpp(
-            UnitMoveTypeLikeCpp::Run,
-            speed,
-        )?)
-    }
-
-    fn apply_represented_forward_speed_adjustments_like_cpp(
-        &self,
-        move_type: UnitMoveTypeLikeCpp,
-        mut speed: f32,
-    ) -> Option<f32> {
-        let normal_speed_cap = self.max_represented_aura_amount_like_cpp(
-            RepresentedAuraEffectLikeCpp::UseNormalMovementSpeed,
-        )? as f32
-            / PLAYER_BASE_MOVE_SPEED_LIKE_CPP[move_type.index()];
-        if normal_speed_cap > 0.0 && speed > normal_speed_cap {
-            speed = normal_speed_cap;
-        }
-        if move_type == UnitMoveTypeLikeCpp::Run {
-            let minimum_speed_rate = self.max_represented_aura_amount_like_cpp(
-                RepresentedAuraEffectLikeCpp::MinimumSpeedRate,
-            )? as f32
-                / PLAYER_BASE_MOVE_SPEED_LIKE_CPP[move_type.index()];
-            if speed < minimum_speed_rate {
-                speed = minimum_speed_rate;
-            }
-        }
-        let slow = self.max_negative_represented_aura_amount_like_cpp(
-            RepresentedAuraEffectLikeCpp::DecreaseSpeed,
-        )?;
-        if slow < 0 {
-            speed *= 1.0 + slow as f32 / 100.0;
-        }
-        let minimum_speed = self
-            .max_represented_aura_amount_like_cpp(RepresentedAuraEffectLikeCpp::MinimumSpeed)?
-            as f32
-            / 100.0;
-        if speed < minimum_speed {
-            speed = minimum_speed;
-        }
-        Some(speed)
-    }
-
-    fn represented_flight_speed_rate_like_cpp(&self) -> Option<f32> {
-        let mounted = self.resolved_player_mounted_like_cpp()?;
-        let (flight_mod, flight_always) = if mounted {
-            (
-                self.max_represented_aura_amount_like_cpp(
-                    RepresentedAuraEffectLikeCpp::MountedFlightSpeed,
-                )?,
-                self.total_represented_aura_amount_multiplier_like_cpp(
-                    RepresentedAuraEffectLikeCpp::MountedFlightSpeedAlways,
-                )?,
-            )
-        } else {
-            (
-                self.total_represented_aura_amount_like_cpp(
-                    RepresentedAuraEffectLikeCpp::FlightSpeed,
-                )? + self.total_represented_aura_amount_like_cpp(
-                    RepresentedAuraEffectLikeCpp::VehicleFlightSpeed,
-                )?,
-                1.0,
-            )
-        };
-        let flight_not_stack = 1.0
-            + self
-                .max_represented_aura_amount_like_cpp(
-                    RepresentedAuraEffectLikeCpp::FlightSpeedNotStack,
-                )?
-                .max(0) as f32
-                / 100.0;
-
-        let speed = flight_always.max(flight_not_stack) * (1.0 + flight_mod.max(0) as f32 / 100.0);
-        Some(self.apply_represented_forward_speed_adjustments_like_cpp(
-            UnitMoveTypeLikeCpp::Flight,
-            speed,
-        )?)
-    }
-
-    fn represented_swim_speed_rate_like_cpp(&self) -> Option<f32> {
-        let speed = 1.0
-            + self
-                .max_represented_aura_amount_like_cpp(RepresentedAuraEffectLikeCpp::SwimSpeed)?
-                .max(0) as f32
-                / 100.0;
-        self.apply_represented_forward_speed_adjustments_like_cpp(UnitMoveTypeLikeCpp::Swim, speed)
-    }
-
-    fn represented_backward_speed_rate_like_cpp(&self) -> Option<f32> {
-        let mut speed = 1.0;
-        let slow = self.max_negative_represented_aura_amount_like_cpp(
-            RepresentedAuraEffectLikeCpp::DecreaseSpeed,
-        )?;
-        if slow < 0 {
-            speed *= 1.0 + slow as f32 / 100.0;
-        }
-        let minimum_speed = self
-            .max_represented_aura_amount_like_cpp(RepresentedAuraEffectLikeCpp::MinimumSpeed)?
-            as f32
-            / 100.0;
-        if speed < minimum_speed {
-            speed = minimum_speed;
-        }
-        Some(speed)
-    }
-
-    fn recompute_represented_run_speed_rate_like_cpp(&mut self) {
-        let Some(speed) = self.represented_run_speed_rate_like_cpp() else {
-            return;
-        };
-        self.set_player_movement_speed_rate_and_notify_like_cpp(UnitMoveTypeLikeCpp::Run, speed);
-    }
-
-    fn recompute_represented_flight_speed_rate_like_cpp(&mut self) {
-        let Some(speed) = self.represented_flight_speed_rate_like_cpp() else {
-            return;
-        };
-        self.set_player_movement_speed_rate_and_notify_like_cpp(UnitMoveTypeLikeCpp::Flight, speed);
-    }
-
-    fn recompute_represented_swim_speed_rate_like_cpp(&mut self) {
-        let Some(speed) = self.represented_swim_speed_rate_like_cpp() else {
-            return;
-        };
-        self.set_player_movement_speed_rate_and_notify_like_cpp(UnitMoveTypeLikeCpp::Swim, speed);
-    }
-
-    fn recompute_represented_backward_speed_rates_like_cpp(&mut self) {
-        let Some(speed) = self.represented_backward_speed_rate_like_cpp() else {
-            return;
-        };
-        self.set_player_movement_speed_rate_and_notify_like_cpp(
-            UnitMoveTypeLikeCpp::RunBack,
-            speed,
-        );
-        self.set_player_movement_speed_rate_and_notify_like_cpp(
-            UnitMoveTypeLikeCpp::SwimBack,
-            speed,
-        );
-        self.set_player_movement_speed_rate_and_notify_like_cpp(
-            UnitMoveTypeLikeCpp::FlightBack,
-            speed,
-        );
-    }
-
-    fn recompute_represented_mounted_speed_rates_like_cpp(&mut self) {
-        self.recompute_represented_run_speed_rate_like_cpp();
-        self.recompute_represented_flight_speed_rate_like_cpp();
-    }
-
-    fn recompute_represented_forward_speed_rates_like_cpp(&mut self) {
-        self.recompute_represented_run_speed_rate_like_cpp();
-        self.recompute_represented_swim_speed_rate_like_cpp();
-        self.recompute_represented_flight_speed_rate_like_cpp();
-    }
-
-    fn player_movement_speed_opcodes_like_cpp(
-        move_type: UnitMoveTypeLikeCpp,
-    ) -> Option<(ServerOpcodes, ServerOpcodes)> {
-        match move_type {
-            UnitMoveTypeLikeCpp::Run => Some((
-                ServerOpcodes::MoveSetRunSpeed,
-                ServerOpcodes::MoveUpdateRunSpeed,
-            )),
-            UnitMoveTypeLikeCpp::Flight => Some((
-                ServerOpcodes::MoveSetFlightSpeed,
-                ServerOpcodes::MoveUpdateFlightSpeed,
-            )),
-            UnitMoveTypeLikeCpp::Swim => Some((
-                ServerOpcodes::MoveSetSwimSpeed,
-                ServerOpcodes::MoveUpdateSwimSpeed,
-            )),
-            UnitMoveTypeLikeCpp::RunBack => Some((
-                ServerOpcodes::MoveSetRunBackSpeed,
-                ServerOpcodes::MoveUpdateRunBackSpeed,
-            )),
-            UnitMoveTypeLikeCpp::SwimBack => Some((
-                ServerOpcodes::MoveSetSwimBackSpeed,
-                ServerOpcodes::MoveUpdateSwimBackSpeed,
-            )),
-            UnitMoveTypeLikeCpp::FlightBack => Some((
-                ServerOpcodes::MoveSetFlightBackSpeed,
-                ServerOpcodes::MoveUpdateFlightBackSpeed,
-            )),
-            _ => None,
-        }
     }
 
     fn represented_pet_position_like_cpp(&self, pet_guid: ObjectGuid) -> Option<Position> {
@@ -39185,35 +36490,6 @@ impl WorldSession {
         }
     }
 
-    fn send_player_move_set_flag_like_cpp(&mut self, opcode: ServerOpcodes) {
-        use wow_packet::ServerPacket;
-
-        let Some(player_guid) = self.player_guid() else {
-            return;
-        };
-        let Some(sequence_index) = self.next_movement_counter_like_cpp() else {
-            return;
-        };
-
-        let self_packet = wow_packet::packets::movement::MoveSetFlag {
-            opcode,
-            mover_guid: player_guid,
-            sequence_index,
-        }
-        .to_bytes();
-        if self.send_tx().send(self_packet).is_err() {
-            warn!("Send channel closed for account {}", self.account_id);
-        }
-
-        let Some(status) = self.current_player_movement_info_like_cpp(player_guid) else {
-            return;
-        };
-        self.broadcast_to_movement_set_like_cpp(
-            wow_packet::packets::movement::MoveUpdate { info: status }.to_bytes(),
-            false,
-        );
-    }
-
     fn set_represented_can_fly_like_cpp(&mut self, enable: bool) -> bool {
         let Some(mut movement_flags) = self.resolved_player_movement_flags_like_cpp() else {
             return false;
@@ -39239,22 +36515,6 @@ impl WorldSession {
         } else {
             ServerOpcodes::MoveUnsetCanFly
         });
-        true
-    }
-
-    fn move_represented_player_fall_like_cpp(&mut self) -> bool {
-        let Some(mut movement_flags) = self.resolved_player_movement_flags_like_cpp() else {
-            return false;
-        };
-        if movement_flags.contains(MovementFlag::DISABLE_GRAVITY) {
-            return false;
-        }
-
-        movement_flags.insert(MovementFlag::FALLING);
-        self.set_player_movement_flags_like_cpp(movement_flags);
-        if let Some(position) = self.player_position_like_cpp() {
-            self.set_fall_information_like_cpp(0, position.z);
-        }
         true
     }
 
@@ -39354,72 +36614,6 @@ impl WorldSession {
         self.send_represented_pet_spline_speed_like_cpp(pet_guid, move_type, rate);
     }
 
-    fn set_player_movement_speed_rate_and_notify_like_cpp(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-        rate: f32,
-    ) {
-        let rate = rate.max(0.01);
-        let Some(current_rate) = self.resolved_player_movement_speed_rate_like_cpp(move_type)
-        else {
-            return;
-        };
-        if current_rate == rate {
-            return;
-        }
-        if !self.set_player_movement_speed_rate_like_cpp_inner(move_type, rate) {
-            return;
-        }
-        self.propagate_represented_player_speed_to_pet_like_cpp(move_type, rate);
-
-        let Some(player_guid) = self.player_guid() else {
-            return;
-        };
-        let Some((set_opcode, update_opcode)) =
-            Self::player_movement_speed_opcodes_like_cpp(move_type)
-        else {
-            return;
-        };
-
-        if self
-            .increment_forced_speed_changes_like_cpp(move_type)
-            .is_none()
-        {
-            return;
-        }
-
-        let Some(speed) = self.resolved_player_movement_speed_like_cpp(move_type) else {
-            return;
-        };
-        let Some(sequence_index) = self.next_movement_counter_like_cpp() else {
-            return;
-        };
-
-        let self_packet = wow_packet::packets::movement::MoveSetSpeed {
-            opcode: set_opcode,
-            mover_guid: player_guid,
-            sequence_index,
-            speed,
-        }
-        .to_bytes();
-        if self.send_tx().send(self_packet).is_err() {
-            warn!("Send channel closed for account {}", self.account_id);
-        }
-
-        let Some(status) = self.current_player_movement_info_like_cpp(player_guid) else {
-            return;
-        };
-        self.broadcast_to_movement_set_like_cpp(
-            wow_packet::packets::movement::MoveUpdateSpeed {
-                opcode: update_opcode,
-                status,
-                speed,
-            }
-            .to_bytes(),
-            false,
-        );
-    }
-
     pub(crate) fn trace_anticheat_violation_like_cpp(
         &self,
         rule: &'static str,
@@ -39435,299 +36629,6 @@ impl WorldSession {
             severity,
             "anticheat.violation"
         );
-    }
-
-    pub(crate) fn handle_force_speed_change_ack_like_cpp(
-        &mut self,
-        opcode: ClientOpcodes,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-        speed: f32,
-    ) -> bool {
-        let Some(move_type) = Self::movement_speed_ack_move_type_like_cpp(opcode) else {
-            self.trace_anticheat_violation_like_cpp(
-                "HandleForceSpeedChangeAck.UnknownMoveType",
-                Some(opcode),
-                "kick",
-            );
-            self.record_movement_speed_ack_event_like_cpp(MovementSpeedAckEventLikeCpp {
-                opcode,
-                move_type: None,
-                ack_speed: speed,
-                expected_speed: None,
-                remaining_forced_changes: None,
-                action: MovementSpeedAckActionLikeCpp::Kicked,
-            });
-            return false;
-        };
-
-        if !self.record_validated_movement_ack_like_cpp(opcode, ack, Some(speed)) {
-            self.trace_anticheat_violation_like_cpp(
-                "HandleForceSpeedChangeAck.InvalidMovementAck",
-                Some(opcode),
-                "kick",
-            );
-            self.record_movement_speed_ack_event_like_cpp(MovementSpeedAckEventLikeCpp {
-                opcode,
-                move_type: Some(move_type),
-                ack_speed: speed,
-                expected_speed: None,
-                remaining_forced_changes: None,
-                action: MovementSpeedAckActionLikeCpp::Kicked,
-            });
-            return false;
-        }
-
-        let Some(mut remaining_forced_changes) =
-            self.resolved_forced_speed_changes_like_cpp(move_type)
-        else {
-            return false;
-        };
-        if remaining_forced_changes > 0 {
-            let Some(remaining) = self.consume_forced_speed_change_like_cpp(move_type) else {
-                return false;
-            };
-            remaining_forced_changes = remaining;
-            if remaining_forced_changes > 0 {
-                self.record_movement_speed_ack_event_like_cpp(MovementSpeedAckEventLikeCpp {
-                    opcode,
-                    move_type: Some(move_type),
-                    ack_speed: speed,
-                    expected_speed: self.resolved_player_movement_speed_like_cpp(move_type),
-                    remaining_forced_changes: Some(remaining_forced_changes),
-                    action: MovementSpeedAckActionLikeCpp::SkippedPending,
-                });
-                return true;
-            }
-        }
-
-        let Some(expected_speed) = self.resolved_player_movement_speed_like_cpp(move_type) else {
-            return false;
-        };
-        let Some(player_on_transport) = self.player_on_transport_state_like_cpp() else {
-            return false;
-        };
-        let action = if !player_on_transport && (expected_speed - speed).abs() > 0.01 {
-            if expected_speed > speed {
-                // C++ calls SetSpeedRate(GetSpeedRate()) to force the client back to the server value.
-                self.trace_anticheat_violation_like_cpp(
-                    "HandleForceSpeedChangeAck.ClientSpeedLower",
-                    Some(opcode),
-                    "correct",
-                );
-                MovementSpeedAckActionLikeCpp::Corrected
-            } else {
-                self.trace_anticheat_violation_like_cpp(
-                    "HandleForceSpeedChangeAck.ClientSpeedHigher",
-                    Some(opcode),
-                    "kick",
-                );
-                self.kick("WorldSession::HandleForceSpeedChangeAck Incorrect speed");
-                MovementSpeedAckActionLikeCpp::Kicked
-            }
-        } else {
-            MovementSpeedAckActionLikeCpp::Accepted
-        };
-
-        self.record_movement_speed_ack_event_like_cpp(MovementSpeedAckEventLikeCpp {
-            opcode,
-            move_type: Some(move_type),
-            ack_speed: speed,
-            expected_speed: Some(expected_speed),
-            remaining_forced_changes: Some(remaining_forced_changes),
-            action,
-        });
-        !matches!(action, MovementSpeedAckActionLikeCpp::Kicked)
-    }
-
-    pub(crate) fn handle_movement_force_mod_magnitude_ack_like_cpp(
-        &mut self,
-        opcode: ClientOpcodes,
-        ack: &mut wow_packet::packets::movement::MovementAck,
-        speed: f32,
-    ) -> bool {
-        if !self.record_validated_movement_ack_like_cpp(opcode, ack, Some(speed)) {
-            self.trace_anticheat_violation_like_cpp(
-                "HandleMoveSetModMovementForceMagnitudeAck.InvalidMovementAck",
-                Some(opcode),
-                "kick",
-            );
-            self.record_movement_speed_ack_event_like_cpp(MovementSpeedAckEventLikeCpp {
-                opcode,
-                move_type: None,
-                ack_speed: speed,
-                expected_speed: None,
-                remaining_forced_changes: None,
-                action: MovementSpeedAckActionLikeCpp::Kicked,
-            });
-            return false;
-        }
-
-        let Some(mut remaining_forced_changes) =
-            self.resolved_movement_force_mod_magnitude_changes_like_cpp()
-        else {
-            return false;
-        };
-        let Some(expected_magnitude) = self.resolved_movement_force_mod_magnitude_like_cpp() else {
-            return false;
-        };
-        let mut action = MovementSpeedAckActionLikeCpp::Accepted;
-        if remaining_forced_changes > 0 {
-            let Some(remaining) = self.consume_movement_force_mod_magnitude_change_like_cpp()
-            else {
-                return false;
-            };
-            remaining_forced_changes = remaining;
-            if remaining_forced_changes == 0 && (expected_magnitude - speed).abs() > 0.01 {
-                self.trace_anticheat_violation_like_cpp(
-                    "HandleMoveSetModMovementForceMagnitudeAck.IncorrectMagnitude",
-                    Some(opcode),
-                    "kick",
-                );
-                self.kick(
-                    "WorldSession::HandleMoveSetModMovementForceMagnitudeAck Incorrect magnitude",
-                );
-                action = MovementSpeedAckActionLikeCpp::Kicked;
-            }
-        }
-
-        self.record_movement_speed_ack_event_like_cpp(MovementSpeedAckEventLikeCpp {
-            opcode,
-            move_type: None,
-            ack_speed: speed,
-            expected_speed: Some(expected_magnitude),
-            remaining_forced_changes: Some(remaining_forced_changes),
-            action,
-        });
-        !matches!(action, MovementSpeedAckActionLikeCpp::Kicked)
-    }
-
-    fn record_movement_speed_ack_event_like_cpp(&mut self, event: MovementSpeedAckEventLikeCpp) {
-        #[cfg(test)]
-        self.movement_speed_ack_events_like_cpp.push(event);
-        #[cfg(not(test))]
-        let _ = event;
-    }
-
-    fn resolved_forced_speed_changes_like_cpp(&self, move_type: UnitMoveTypeLikeCpp) -> Option<u8> {
-        let index = move_type.index();
-        let canonical = self
-            .with_owned_player_like_cpp(|player| player.forced_speed_changes_like_cpp(index))
-            .flatten();
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.forced_speed_changes_like_cpp[index]);
-        }
-        canonical
-    }
-
-    fn increment_forced_speed_changes_like_cpp(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-    ) -> Option<u8> {
-        let index = move_type.index();
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.increment_forced_speed_changes_like_cpp(index)
-            })
-            .flatten();
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            let count = &mut self.forced_speed_changes_like_cpp[index];
-            *count = count.saturating_add(1);
-            return Some(*count);
-        }
-        canonical
-    }
-
-    fn consume_forced_speed_change_like_cpp(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-    ) -> Option<u8> {
-        let index = move_type.index();
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.consume_forced_speed_change_like_cpp(index)
-            })
-            .flatten();
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            let count = &mut self.forced_speed_changes_like_cpp[index];
-            if *count > 0 {
-                *count = count.saturating_sub(1);
-            }
-            return Some(*count);
-        }
-        canonical
-    }
-
-    fn resolved_movement_force_mod_magnitude_changes_like_cpp(&self) -> Option<u8> {
-        let canonical = self.with_owned_player_like_cpp(|player| {
-            player.movement_force_mod_magnitude_changes_like_cpp()
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.movement_force_mod_magnitude_changes_like_cpp);
-        }
-        canonical
-    }
-
-    fn resolved_movement_force_mod_magnitude_like_cpp(&self) -> Option<f32> {
-        let canonical = self.with_owned_player_like_cpp(|player| {
-            player.unit().movement_force_mod_magnitude_like_cpp()
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(self.movement_force_mod_magnitude_like_cpp);
-        }
-        canonical
-    }
-
-    fn consume_movement_force_mod_magnitude_change_like_cpp(&mut self) -> Option<u8> {
-        let canonical = self.with_owned_player_mut_like_cpp(|player| {
-            player.consume_movement_force_mod_magnitude_change_like_cpp()
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            if self.movement_force_mod_magnitude_changes_like_cpp > 0 {
-                self.movement_force_mod_magnitude_changes_like_cpp = self
-                    .movement_force_mod_magnitude_changes_like_cpp
-                    .saturating_sub(1);
-            }
-            return Some(self.movement_force_mod_magnitude_changes_like_cpp);
-        }
-        canonical
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_forced_speed_changes_like_cpp(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-        count: u8,
-    ) {
-        let index = move_type.index();
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.set_forced_speed_changes_like_cpp(index, count)
-            })
-            .unwrap_or(false);
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.forced_speed_changes_like_cpp[index] = count;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn forced_speed_changes_like_cpp(&self, move_type: UnitMoveTypeLikeCpp) -> u8 {
-        self.resolved_forced_speed_changes_like_cpp(move_type)
-            .expect("test Player forced-speed owner must resolve")
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_player_movement_speed_rate_like_cpp(
-        &mut self,
-        move_type: UnitMoveTypeLikeCpp,
-        rate: f32,
-    ) {
-        let _ = self.set_player_movement_speed_rate_like_cpp_inner(move_type, rate.max(0.01));
     }
 
     #[cfg(test)]
@@ -39802,38 +36703,6 @@ impl WorldSession {
         self.player_transport_state_like_cpp()
             .flatten()
             .map(|state| state.guid)
-    }
-
-    pub(crate) fn set_player_transport_position_like_cpp(&mut self, position: Option<Position>) {
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            if let (Some(state), Some(position)) = (
-                self.player_transport_login_state_like_cpp.as_mut(),
-                position,
-            ) {
-                state.info.x = position.x;
-                state.info.y = position.y;
-                state.info.z = position.z;
-                state.info.o = position.orientation;
-            }
-            return;
-        }
-        let _ = self.with_owned_player_mut_like_cpp(|player| {
-            if let (Some(transport), Some(position)) =
-                (player.gameplay_state_mut().transport.as_mut(), position)
-            {
-                transport.x = position.x;
-                transport.y = position.y;
-                transport.z = position.z;
-                transport.orientation = position.orientation;
-            }
-        });
-    }
-
-    pub(crate) fn player_transport_position_like_cpp(&self) -> Option<Position> {
-        self.player_transport_state_like_cpp()
-            .flatten()
-            .map(|state| Position::new(state.x, state.y, state.z, state.orientation))
     }
 
     pub(crate) fn player_transport_info_like_cpp(
@@ -39918,37 +36787,6 @@ impl WorldSession {
     #[cfg(test)]
     pub(crate) fn set_player_on_transport_like_cpp(&mut self, on_transport: bool) {
         self.player_on_transport_like_cpp = on_transport;
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_movement_force_mod_magnitude_changes_like_cpp(&mut self, count: u8) {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.set_movement_force_mod_magnitude_changes_like_cpp(count);
-            })
-            .is_some();
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.movement_force_mod_magnitude_changes_like_cpp = count;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_movement_force_mod_magnitude_like_cpp(&mut self, magnitude: f32) {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player
-                    .unit_mut()
-                    .set_movement_force_mod_magnitude_like_cpp(magnitude);
-            })
-            .is_some();
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.movement_force_mod_magnitude_like_cpp = magnitude;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn movement_speed_ack_events_like_cpp(&self) -> &[MovementSpeedAckEventLikeCpp] {
-        &self.movement_speed_ack_events_like_cpp
     }
 
     pub(crate) fn destroy_represented_totem_like_cpp(
@@ -40712,38 +37550,6 @@ impl WorldSession {
         sent
     }
 
-    pub(crate) fn send_represented_capture_point_removed_from_last_update_like_cpp(
-        &mut self,
-    ) -> usize {
-        let Some((map_id, instance_id, update_generation, removable_guids)) = self
-            .visible_gameobject_guids_from_last_update_summary_like_cpp(|summary| {
-                summary
-                    .generic_capture_point_removed_guids
-                    .as_slice()
-                    .to_vec()
-            })
-        else {
-            return 0;
-        };
-
-        let mut seen = std::collections::HashSet::new();
-        let mut sent = 0;
-        for guid in removable_guids {
-            if !seen.insert(guid) || !self.client_visible_guids_like_cpp.contains(&guid) {
-                continue;
-            }
-            if !self
-                .represented_capture_point_removed_delivered_like_cpp
-                .insert((map_id, instance_id, update_generation, guid))
-            {
-                continue;
-            }
-            self.send_represented_capture_point_removed_like_cpp(guid);
-            sent += 1;
-        }
-        sent
-    }
-
     fn represented_player_has_active_vehicle_like_cpp(&self) -> bool {
         self.player_mount_vehicle_kit_snapshot_like_cpp()
             .flatten()
@@ -40841,39 +37647,6 @@ impl WorldSession {
             AccessorObjectKind::Pet,
             AccessorObjectKind::DynamicObject,
         ]
-    }
-
-    pub(crate) fn represented_visibility_source_position_like_cpp(&self) -> Option<Position> {
-        let player_position = self.player_position_like_cpp()?;
-        let Some(player_guid) = self.player_guid() else {
-            return Some(player_position);
-        };
-        let Some(seer_guid) = self.represented_seer_guid_like_cpp else {
-            return Some(player_position);
-        };
-        if seer_guid.is_empty() || seer_guid == player_guid {
-            return Some(player_position);
-        }
-
-        let Some(key) = self.current_canonical_player_map_key_like_cpp() else {
-            return Some(player_position);
-        };
-        let Some(manager) = self.canonical_map_manager.as_ref() else {
-            return Some(player_position);
-        };
-        let Ok(manager) = manager.lock() else {
-            return Some(player_position);
-        };
-        manager
-            .find_map(key.map_id, key.instance_id)
-            .and_then(|managed| {
-                managed.map().with_world_object_by_kinds_like_cpp(
-                    seer_guid,
-                    Self::represented_seer_kinds_like_cpp(),
-                    |object| object.position(),
-                )
-            })
-            .or(Some(player_position))
     }
 
     fn visibility_distance_allows_like_cpp(
